refactor(R2): split MorphoDepotLogic into 10 domain mixins
gh/git, control-plane, object-store, RepoClerk, deps, repo-load, contribute,
release, accession, search. MorphoDepotLogic keeps __init__ + class attrs and
inherits them. MorphoDepot.py 6361 -> 3844 lines. Net 12/12 (added Logic-mixin touch).
