+++ MorphoDepot/MorphoDepot.py
@@ -41,6 +41,16 @@ from MorphoDepotLib.forms import (FormBaseQuestion, FormRadioQuestion, FormCheck
 from MorphoDepotLib.accession_form import MorphoDepotAccessionForm
 from MorphoDepotLib.search_form import MorphoDepotSearchForm
 from MorphoDepotLib.screenshot_dialog import ScreenshotReviewDialog
+from MorphoDepotLib.logic_deps import DepsMixin
+from MorphoDepotLib.logic_github import GitHubMixin
+from MorphoDepotLib.logic_controlplane import ControlPlaneMixin
+from MorphoDepotLib.logic_objectstore import ObjectStoreMixin
+from MorphoDepotLib.logic_repoclerk import RepoClerkMixin
+from MorphoDepotLib.logic_repo import RepoMixin
+from MorphoDepotLib.logic_contribute import ContributeMixin
+from MorphoDepotLib.logic_release import ReleaseMixin
+from MorphoDepotLib.logic_accession import AccessionMixin
+from MorphoDepotLib.logic_search import SearchMixin
 
 
 class MorphoDepot(ScriptedLoadableModule):
@@ -3320,7 +3330,7 @@ class MorphoDepotWidget(ScriptedLoadableModuleWidget, VTKObservationMixin, Enabl
                                        dontShowAgainSettingsKey = "MorphoDepot/DontShowPreviewNotice")
 
 
-class MorphoDepotLogic(ScriptedLoadableModuleLogic):
+class MorphoDepotLogic(ScriptedLoadableModuleLogic, DepsMixin, GitHubMixin, ControlPlaneMixin, ObjectStoreMixin, RepoClerkMixin, RepoMixin, ContributeMixin, ReleaseMixin, AccessionMixin, SearchMixin):
     """This class should implement all the actual
     computation done by your module.  The interface
     should be such that other python code can import
@@ -3370,2533 +3380,6 @@ class MorphoDepotLogic(ScriptedLoadableModuleLogic):
 
         qt.QSettings().setValue("MorphoDepot/gitPath", self.gitExecutablePath)
         qt.QSettings().setValue("MorphoDepot/ghPath", self.ghExecutablePath)
-
-    def slicerVersionCheck(self):
-        return hasattr(slicer.vtkSegment, "SetTerminology")
-
-    def resolveVolumeURL(self, volumeRef, repoNameWithOwner):
-        """Convert a source_volume file reference to a full download URL.
-        Accepts both legacy full URLs (backwards compatible) and new relative paths.
-        A source_volume now holds an absolute object-store (JS2) URL; because it starts with
-        "http" it passes straight through here unchanged — only the older relative
-        "releases/download/v1/..." pointers are re-based onto the repo owner.
-        """
-        if volumeRef.startswith("http"):
-            return volumeRef  # full URL (object-store or legacy hardcoded) — use as-is
-        return f"https://github.com/{repoNameWithOwner}/{volumeRef}"
-
-    def uploadSourceVolumeToObjectStore(self, sourceFilePath, sha256, creator, repo, filename):
-        """Upload the source volume to the MorphoDepot object store (JS2) with a server-mediated
-        S3 MULTIPART upload, and return the public URL of the stored object.
-
-        The signing service holds the bucket credentials; the client never does.  The server
-        runs create / complete / abort and signs each part URL on demand, so a slow upload never
-        outlives a part URL (the client re-signs on a 403).  The client PUTs each chunk directly
-        to S3 and verifies the returned part ETag.  On any failure the whole upload is aborted so
-        no orphaned parts linger (a bucket lifecycle rule is the backstop).  The object is keyed
-        {creator}/{repo}/{filename} with the volume's identity (sha256, creator, repo, original
-        filename) stamped as immutable S3 user-metadata at create time; integrity is guaranteed
-        end-to-end by the committed source_volume_checksum (SHA-256 of the whole file).  Endpoint
-        + fallback token come from QSettings (MorphoDepot/uploadSignEndpoint,
-        MorphoDepot/uploadSignToken).  See docs/ObjectStorage-model.md."""
-        import requests
-        qsettings = qt.QSettings()
-        signEndpoint = qsettings.value("MorphoDepot/uploadSignEndpoint",
-                                       "https://join.morphodepot.org/uploads/sign")
-        # Derive the multipart base: …/uploads/sign -> …/uploads/multipart
-        base = signEndpoint.rsplit("/uploads/", 1)[0] + "/uploads/multipart"
-        # Authenticate with the user's own gh token (member tier — "git + gh, nothing else").
-        # Falls back to a QSettings shared token only if gh is somehow unavailable.
-        try:
-            token = self._ghToken()
-        except Exception:
-            token = ""
-        if not token:
-            token = qsettings.value("MorphoDepot/uploadSignToken", "")
-        authHeaders = {"Authorization": f"Bearer {token}"} if token else {}
-
-        def post(path, body):
-            r = requests.post(f"{base}/{path}", json=body, headers=authHeaders, timeout=60)
-            if r.status_code != 200:
-                raise RuntimeError(f"Object-store multipart {path} failed "
-                                   f"({r.status_code}): {r.text}")
-            return r.json()
-
-        size = os.path.getsize(sourceFilePath)
-        self.progressMethod("Requesting multipart upload from the object-store signing service...")
-        created = post("create", {"sha256": sha256, "size": size, "creator": creator,
-                                  "repo": repo, "filename": filename})
-        publicURL = created["public_url"]
-        if created.get("already_exists"):
-            self.progressMethod("Source volume already in the object store; skipping upload.")
-            return publicURL
-
-        key = created["key"]
-        uploadId = created["upload_id"]
-        partSize = int(created.get("part_size") or (128 * 2**20))
-        parts = []
-        try:
-            with open(sourceFilePath, "rb") as fp:
-                partNumber = 0
-                uploaded = 0
-                while True:
-                    chunk = fp.read(partSize)
-                    if not chunk:
-                        break
-                    partNumber += 1
-                    etag = self._uploadOneVolumePart(post, key, uploadId, partNumber, chunk)
-                    parts.append({"part_number": partNumber, "etag": etag})
-                    uploaded += len(chunk)
-                    self.progressMethod(f"Uploaded {uploaded / 2**20:.0f} / {size / 2**20:.0f} MB "
-                                        "to the object store...")
-            self.progressMethod("Finalizing multipart upload...")
-            completed = post("complete", {"key": key, "upload_id": uploadId, "parts": parts})
-            return completed.get("public_url", publicURL)
-        except Exception:
-            # Abort so a partial upload leaves no orphaned parts (lifecycle rule is the backstop).
-            try:
-                post("abort", {"key": key, "upload_id": uploadId})
-                self.progressMethod("Upload failed — aborted the multipart upload (no orphaned data).")
-            except Exception as abortError:
-                logging.warning(f"Could not abort multipart upload {key}/{uploadId}: {abortError}")
-            raise
-
-    def _uploadOneVolumePart(self, post, key, uploadId, partNumber, chunk):
-        """PUT one multipart chunk to a freshly-signed part URL, verify its ETag, and return it.
-        Re-signs the URL (handling expiry) and retries on transient failure."""
-        import requests
-        import hashlib
-        expectedMd5 = hashlib.md5(chunk).hexdigest()
-        lastError = None
-        for _attempt in range(4):
-            # Send the exact chunk length so the signing service binds it into the part URL
-            # (S3 then rejects a mismatched body) — part of the server-side per-file size cap.
-            signed = post("sign", {"key": key, "upload_id": uploadId, "part_number": partNumber,
-                                   "content_length": len(chunk)})
-            try:
-                resp = requests.put(signed["url"], data=chunk, timeout=None)
-            except Exception as e:
-                lastError = e
-                continue
-            if resp.status_code == 403:   # part URL expired — re-sign and retry
-                lastError = RuntimeError("part URL expired (403)")
-                continue
-            if resp.status_code not in (200, 201):
-                lastError = RuntimeError(f"part {partNumber} PUT failed "
-                                         f"({resp.status_code}): {resp.text}")
-                continue
-            etag = (resp.headers.get("ETag") or "").strip()
-            clean = etag.strip('"').lower()
-            # For our unencrypted bucket the part ETag is the part's MD5 — verify when it looks
-            # like one (skip for non-MD5 ETags; the whole-file SHA-256 still backstops on download).
-            if len(clean) == 32 and all(c in "0123456789abcdef" for c in clean) and clean != expectedMd5:
-                lastError = RuntimeError(f"part {partNumber} checksum mismatch")
-                continue
-            if not etag:
-                lastError = RuntimeError(f"part {partNumber} returned no ETag")
-                continue
-            return etag
-        raise RuntimeError(f"part {partNumber} failed after retries: {lastError}")
-
-    # --- Membership tier + App control plane (member repos are born in-org, App-mediated) ---
-
-    morphoDepotOrg = "MorphoDepot"
-
-    # Throwaway org for the developer Reload-and-Test self-test.  Test repos are created here
-    # directly (the developer's own gh rights — no App, no S3), never on a personal account or in
-    # the production org, and are deleted at the end of the test.  Both dev accounts are members.
-    morphoDepotTestingOrg = "MorphoDepotTesting"
-
-    def controlPlaneBase(self):
-        return qt.QSettings().value(
-            "MorphoDepot/controlPlaneBase", "https://join.morphodepot.org").rstrip("/")
-
-    def orgMembershipStatus(self, org=None):
-        """Tri-state membership against the MorphoDepot org via the App control plane (`/me`):
-        'member' | 'non_member' | 'unknown' (could not determine — App unreachable / invalid token).
-        The App verifies membership with its OWN token, so the user's gh token needs no org scope.
-        Caches only a CONFIRMED result; 'unknown' is never trusted from cache, so a transient failure
-        re-checks (and never reports a real member's outage as a membership problem)."""
-        org = org or self.morphoDepotOrg
-        cache = getattr(self, "_orgMemberCache", None)
-        if cache is not None and cache[0] == org and cache[1] in ("member", "non_member"):
-            return cache[1]
-        try:
-            info = self.controlPlaneRequest("me", {})
-            status = "member" if bool(info.get("is_member")) else "non_member"
-        except Exception as e:
-            # Do NOT cache 'unknown' (it would also evict a prior confirmed result) — a transient
-            # failure must re-check on the next call rather than stick.
-            logging.warning(f"Membership check failed (status unknown): {e}")
-            return "unknown"
-        self._orgMemberCache = (org, status)
-        return status
-
-    def userIsOrgMember(self, org=None):
-        """True only when membership is CONFIRMED.  Boolean callers treat 'unknown' as non-member;
-        the create path uses orgMembershipStatus() directly to distinguish 'not a member' from
-        'could not verify' so a transient outage is never reported as a membership problem.  (Asks
-        the App `/me`, which checks membership with the App token — the user's gh token needs no
-        org scope.  Reload the module after switching gh accounts or joining the org.)"""
-        return self.orgMembershipStatus(org) == "member"
-
-    def _ghToken(self):
-        """The user's GitHub token (from gh) used to authenticate to the App control plane."""
-        import subprocess
-        try:
-            out = subprocess.run([self.ghExecutablePath, "auth", "token"],
-                                 capture_output=True, text=True, timeout=15)
-            return (out.stdout or "").strip()
-        except Exception as e:
-            raise RuntimeError(f"Could not get a GitHub token from gh: {e}")
-
-    def controlPlaneRequest(self, path, body):
-        """POST to the intake App control plane, authenticated by the user's gh token.
-        Returns the parsed JSON, or raises with the server's error detail."""
-        import requests
-        token = self._ghToken()
-        if not token:
-            raise RuntimeError("Not signed in to GitHub — run `gh auth login` first.")
-        r = requests.post(f"{self.controlPlaneBase()}/{path}", json=body,
-                          headers={"Authorization": f"Bearer {token}"}, timeout=120)
-        if r.status_code != 200:
-            try:
-                detail = r.json().get("detail", r.text)
-            except Exception:
-                detail = r.text
-            raise RuntimeError(f"Control plane '{path}' failed ({r.status_code}): {detail}")
-        return r.json()
-
-    def volumeChecksumIndexURL(self):
-        """RepoClerk's published checksum->repo index (GitHub Pages JSON)."""
-        return qt.QSettings().value(
-            "MorphoDepot/volumeChecksumIndexURL",
-            "https://MorphoDepot.github.io/RepoClerk/volume-checksums.json")
-
-    def duplicateVolumeRepos(self, checksum, exclude=None):
-        """Repos (nameWithOwner) that already hold a volume with this SHA-256, per RepoClerk's
-        published checksum->repo index.  Best-effort and ADVISORY: returns [] on any failure
-        (network down, index missing/stale) so it never blocks staging or publishing.  The index
-        lags RepoClerk's crawl (~6 h), so a very recently published duplicate may not appear yet.
-        `exclude` drops the repo being created/published (a repo never duplicates itself)."""
-        if not checksum:
-            return []
-        sha = checksum.strip()
-        if ":" in sha:  # the committed file is "SHA256:<hex>"; the index stores bare hex
-            sha = sha.split(":", 1)[1].strip()
-        sha = sha.lower()
-        if not sha:
-            return []
-        try:
-            resp = requests.get(self.volumeChecksumIndexURL(), timeout=10)
-            resp.raise_for_status()
-            index = resp.json().get("checksums", {})
-        except Exception as e:
-            logging.warning(f"Duplicate-volume check skipped ({e})")
-            return []
-        return [r for r in index.get(sha, []) if r != exclude]
-
-    def localRepositoryDirectory(self):
-        repoDirectory = os.path.normpath(slicer.util.settingsValue("MorphoDepot/repoDirectory", "") or "")
-        if repoDirectory == "" or repoDirectory == ".":
-            defaultScenePath = os.path.normpath(slicer.app.defaultScenePath)
-            defaultRepoDir = os.path.join(defaultScenePath, "MorphoDepot")
-            self.setLocalRepositoryDirectory(defaultRepoDir)
-            repoDirectory = defaultRepoDir
-
-        if repoDirectory and not os.path.exists(repoDirectory):
-            message = f"The repository directory does not exist:\n\n{repoDirectory}\n\nCreate it now?"
-            if slicer.util.confirmOkCancelDisplay(message, windowTitle="Create Directory"):
-                try:
-                    os.makedirs(repoDirectory)
-                except OSError as e:
-                    logging.error(f"Could not create repository directory {repoDirectory}: {e}")
-
-        return repoDirectory
-
-    def setLocalRepositoryDirectory(self, repoDir):
-        qt.QSettings().setValue("MorphoDepot/repoDirectory", repoDir)
-
-    def checkPythonDependencies(self):
-        """See if pygbif and idigbio are available.
-        The GitPython package is installed by default in slicer.
-        """
-        try:
-            import pygbif
-        except ModuleNotFoundError:
-            return False
-
-        try:
-            import idigbio
-        except ModuleNotFoundError:
-            return False
-
-        return True
-
-    def installPythonDependencies(self):
-        """Install pygbif and idigbio if needed
-        """
-        try:
-            import pygbif
-        except ModuleNotFoundError:
-            self.progressMethod(f"Installing pygbif")
-            slicer.util.pip_install("pygbif")
-            import pygbif
-
-        try:
-            import idigbio
-        except ModuleNotFoundError:
-            self.progressMethod(f"Installing idigbio")
-            slicer.util.pip_install("idigbio")
-            import idigbio
-
-    def checkCommand(self, command):
-        try:
-            completedProcess = subprocess.run(command, capture_output=True)
-            returnCode = completedProcess.returncode
-            stdout = completedProcess.stdout
-            stderr = completedProcess.stderr
-        except Exception as e:
-            stdout =  ""
-            stderr = str(e)
-            returnCode = -1
-        if returnCode != 0:
-            self.progressMethod(f"{command} failed to run, returned {returnCode}")
-            self.progressMethod(stdout)
-            self.progressMethod(stderr)
-            return False
-        return True
-
-    def checkGitDependencies(self):
-        """Check that git, and gh are available
-        """
-        if not (self.gitExecutablePath and self.ghExecutablePath):
-            self.progressMethod("git/gh paths are not set")
-            return False
-        if not (os.path.exists(self.gitExecutablePath) and os.path.exists(self.ghExecutablePath)):
-            self.progressMethod("bad git/gh paths")
-            self.progressMethod(f"git path is {self.gitExecutablePath}")
-            self.progressMethod(f"gh path is {self.ghExecutablePath}")
-            return False
-        if not self.checkCommand([self.gitExecutablePath, '--version']):
-            return False
-        if not self.checkCommand([self.ghExecutablePath, 'auth', 'status']):
-            return False
-        return True
-
-    def gh(self, command):
-        """Execute `gh` command.  Multiline input string accepted for readablity.
-        Do not include `gh` in the command string"""
-        if not self.ghExecutablePath or self.ghExecutablePath == "":
-            logging.error("Error, gh not found")
-            return "Error, gh not found"
-        if command.__class__() == "":
-            commandList = command.replace("\n", " ").split()
-        elif command.__class__() == []:
-            commandList = command
-        else:
-            raise TypeError("gh command must be a string or list")
-        self.progressMethod(" ".join(commandList))
-        fullCommandList = [self.ghExecutablePath] + commandList
-
-        baseDelay = 1
-        attempts = 4
-        for attempt in range(attempts):
-            originalLocale = locale.setlocale(locale.LC_ALL)
-            locale.setlocale(locale.LC_ALL, "en_US.UTF-8")
-            process = slicer.util.launchConsoleProcess(fullCommandList)
-            result = process.communicate()
-            locale.setlocale(locale.LC_ALL, originalLocale)
-            needRetry = result[0].find("error: 503") != -1
-            if process.returncode == 0 or not needRetry:
-                if attempt > 0:
-                    print(f"Command succeeded after try {attempt}")
-                break
-            error_message = f"gh command failed: {' '.join(commandList)}\nCode: {process.returncode}, Output: {result}"
-            print(error_message)
-            delay = baseDelay * (2 ** attempt)
-            self.progressMethod(f"gh returned {process.returncode}, sleeping {delay} seconds before retry")
-            time.sleep(delay)
-        if process.returncode != 0:
-            error_message = f"gh command failed: {' '.join(commandList)}\nOutput: {result}"
-            logging.error(error_message)
-            self.progressMethod(f"gh command error: {result}")
-            raise RuntimeError(error_message)
-        self.progressMethod(f"gh command finished: {result}")
-        return result[0]
-
-    def getGitConfig(self, key):
-        """Get a value from the global git config."""
-        if not self.gitExecutablePath:
-            return ""
-        try:
-            command = [self.gitExecutablePath, 'config', '--global', key]
-            result = subprocess.check_output(command, universal_newlines=True).strip()
-            return result
-        except (subprocess.CalledProcessError, FileNotFoundError):
-            return ""
-
-    def setGitConfig(self, key, value):
-        """Set a value in the global git config."""
-        if not self.gitExecutablePath:
-            return
-        try:
-            subprocess.check_call([self.gitExecutablePath, 'config', '--global', key, value])
-        except (subprocess.CalledProcessError, FileNotFoundError) as e:
-            logging.error(f"Failed to set git config {key}: {e}")
-
-    def ghJSON(self, command):
-        """Wrapper around gh that returns json loaded data or an empty list on error"""
-        jsonString = self.gh(command)
-        if jsonString:
-            try:
-                return json.loads(jsonString)
-            except (ValueError, json.JSONDecodeError):
-                logging.warning(f"ghJSON: could not parse gh output for {command!r}")
-        return []
-
-    def hasWorkflowScope(self):
-        """True if the active gh token carries the `workflow` OAuth scope, which GitHub requires
-        to push or create files under `.github/workflows/`.  Used to gate the optional auto-assign
-        workflow at repo creation: without it, a push that includes a workflow file is rejected,
-        so the option is offered only when the scope is present.  Reads GitHub's authoritative
-        `X-Oauth-Scopes` response header via `gh api -i`."""
-        try:
-            output = self.gh("api -i user")
-        except Exception as e:
-            logging.warning(f"Could not determine gh token scopes: {e}")
-            return False
-        for line in output.splitlines():
-            if line.lower().startswith("x-oauth-scopes:"):
-                scopes = [s.strip() for s in line.split(":", 1)[1].split(",")]
-                return "workflow" in scopes
-        return False
-
-    def ghTopicClearCache(self):
-        self.gh("config clear-cache")
-
-    REPOCLERK_URL = "https://github.com/MorphoDepot/RepoClerk"
-    REPOCLERK_SIZE_LIMIT_MB = 100
-
-    def refreshRepoClerk(self):
-        """Maintain a shallow clone of RepoClerk and pull latest journals.
-        Returns the clone path on success, or None on failure (triggers fallback to direct API)."""
-        clonePath = os.path.join(self.localRepositoryDirectory(), "MorphoDepotCaches", "RepoClerk")
-        try:
-            if os.path.exists(clonePath):
-                total = sum(
-                    os.path.getsize(os.path.join(root, f))
-                    for root, _, files in os.walk(clonePath)
-                    for f in files
-                )
-                if total / 1e6 > self.REPOCLERK_SIZE_LIMIT_MB:
-                    shutil.rmtree(clonePath)
-            if not os.path.exists(clonePath):
-                subprocess.run(
-                    [self.gitExecutablePath, 'clone', '--depth', '1', self.REPOCLERK_URL, clonePath],
-                    check=True, capture_output=True
-                )
-            else:
-                subprocess.run(
-                    [self.gitExecutablePath, 'pull'],
-                    cwd=clonePath, check=True, capture_output=True
-                )
-            return clonePath
-        except Exception as e:
-            logging.warning(f"RepoClerk refresh failed: {e}")
-            return None
-
-    def repoClerkJournals(self, clonePath):
-        """Read all journal JSON files from a RepoClerk clone. Returns list of journal dicts."""
-        journals = []
-        journalsDir = os.path.join(clonePath, "journals")
-        if not os.path.exists(journalsDir):
-            return journals
-        for entry in os.scandir(journalsDir):
-            if entry.name.endswith(".json"):
-                try:
-                    with open(entry.path) as f:
-                        journals.append(json.load(f))
-                except Exception as e:
-                    logging.warning(f"Failed to read journal {entry.path}: {e}")
-        return journals
-
-    def notifyRepoClerk(self, nameWithOwner):
-        """Request a RepoClerk journal update by opening an update-request issue.
-        Issue creation works for any authenticated GitHub user (no write access to RepoClerk needed).
-        The drain loop picks up and closes these issues as it processes them."""
-        self.gh([
-            "issue", "create",
-            "--repo", "MorphoDepot/RepoClerk",
-            "--title", f"update {nameWithOwner}",
-            "--label", "update-request",
-            "--body", "",
-        ])
-
-    def hasRepoClerkUpdatePending(self):
-        """Returns True if there are open update-request issues in RepoClerk.
-        Issues are only closed once the drain loop has finished processing them,
-        so this covers the full window from queuing through page rebuild completion."""
-        result = self.gh("issue list --repo MorphoDepot/RepoClerk --state open --label update-request --json number")
-        if result:
-            try:
-                return len(json.loads(result)) > 0
-            except Exception:
-                pass
-        return False
-
-    def _journalsToTopicData(self, journals):
-        """Convert RepoClerk journal list to the dict shape that ghTopicData() returned."""
-        result = []
-        for j in journals:
-            issues_nodes = [
-                {
-                    "number": issue["number"],
-                    "title": issue["title"],
-                    "url": issue["url"],
-                    "author": {"login": issue["author"]} if issue.get("author") else None,
-                    "assignees": {"nodes": [{"login": a} for a in issue.get("assignees", [])]},
-                }
-                for issue in j.get("openIssues", [])
-            ]
-            prs_nodes = [
-                {
-                    "number": pr["number"],
-                    "title": pr["title"],
-                    "isDraft": pr["isDraft"],
-                    "url": pr["url"],
-                    "author": {"login": pr["author"]} if pr.get("author") else None,
-                    "closingIssuesReferences": {
-                        "nodes": [
-                            {
-                                "number": pr["closingIssue"]["number"],
-                                "title": pr["closingIssue"]["title"],
-                                "repository": {"owner": {"login": pr["closingIssue"]["repoOwner"]}},
-                            }
-                        ] if pr.get("closingIssue") else []
-                    },
-                }
-                for pr in j.get("openPRs", [])
-            ]
-            result.append({
-                "nameWithOwner": j["nameWithOwner"],
-                "curator": j.get("curator"),
-                "pullRequests": {"nodes": prs_nodes},
-                "issues": {"nodes": issues_nodes},
-            })
-        return result
-
-    def ghTopicData(self, topic="MorphoDepot"):
-        clonePath = self.refreshRepoClerk()
-        if clonePath:
-            journals = self.repoClerkJournals(clonePath)
-            if journals:
-                return self._journalsToTopicData(journals)
-        logging.warning("RepoClerk journals unavailable — returning empty topic data")
-        return []
-
-    def morphoRepos(self):
-        clonePath = self.refreshRepoClerk()
-        if clonePath:
-            journals = self.repoClerkJournals(clonePath)
-            if journals:
-                result = []
-                for j in journals:
-                    openIssues = j.get("openIssues", []) or []
-                    openPRs = j.get("openPRs", []) or []
-                    # Carry issue/PR totalCount + nodes from the journal so the Release tab's counts,
-                    # announcement targets, and tooltip are accurate (they previously read 0 / empty).
-                    issueNodes = [
-                        {"number": i.get("number"), "title": i.get("title"),
-                         "author": {"login": i.get("author")} if i.get("author") else None,
-                         "assignees": {"nodes": [{"login": a} for a in (i.get("assignees") or [])]}}
-                        for i in openIssues
-                    ]
-                    prNodes = [
-                        {"number": p.get("number"), "title": p.get("title"),
-                         "isDraft": p.get("isDraft"),
-                         "author": {"login": p.get("author")} if p.get("author") else None}
-                        for p in openPRs
-                    ]
-                    result.append({
-                        "name": j["nameWithOwner"].split("/")[1],
-                        "owner": {"login": j["nameWithOwner"].split("/")[0]},
-                        "pushedAt": j.get("pushedAt", ""),
-                        "curator": j.get("curator"),
-                        "issues": {"totalCount": len(openIssues), "nodes": issueNodes},
-                        "pullRequests": {"totalCount": len(openPRs), "nodes": prNodes},
-                    })
-                return result
-        logging.warning("RepoClerk journals unavailable — returning empty repo list")
-        return []
-
-    def whoami(self):
-        """ Get the active gh login.  Parse the '... account <login> ...' token by regex (robust to
-        gh's status-line wording) rather than a fixed word index; fall back to the stable API."""
-        status = self.gh("auth status --active") or ""
-        match = re.search(r"account\s+(\S+)", status)
-        if match:
-            return match.group(1)
-        return self.gh("api user --jq .login").strip()
-
-    def ghUserProfile(self):
-        """Return {login, name, email} for the active gh account.
-
-        `email` is the account's verified **primary** email from `user/emails`, so it is
-        available even when the user keeps their profile email private — this needs the
-        `user:email` scope, which the prerequisites instruct everyone to grant
-        (`gh auth login -s user:email`).  Falls back to the public profile email (`gh api user`)
-        if `user/emails` is unavailable (e.g. the scope was not granted), then to empty.
-        Returns empty strings on any error so callers can fall back to asking the user."""
-        try:
-            data = self.ghJSON(["api", "user"])
-        except Exception as e:
-            logging.warning(f"Could not fetch gh user profile: {e}")
-            return {"login": "", "name": "", "email": ""}
-        if not isinstance(data, dict):
-            return {"login": "", "name": "", "email": ""}
-        # Verified primary email (works for private emails) via the user:email scope.
-        email = ""
-        try:
-            emails = self.ghJSON(["api", "user/emails"])
-            if isinstance(emails, list):
-                primary = next((e for e in emails if isinstance(e, dict) and e.get("primary")), None)
-                if primary:
-                    email = primary.get("email") or ""
-        except Exception as e:
-            logging.warning(f"Could not read user/emails — is the 'user:email' scope granted? ({e})")
-        if not email:
-            email = data.get("email") or ""  # fall back to the public profile email
-        return {
-            "login": data.get("login") or "",
-            "name": data.get("name") or "",
-            "email": email,
-        }
-
-    def userOrganizations(self):
-        """Return the list of GitHub organization logins the active gh user belongs to.
-
-        A user may belong to several organizations; all are returned so the Create tab
-        can offer them as repository destinations.  Returns an empty list on any error
-        (e.g. gh not configured) so callers can fall back to the personal account.
-        """
-        try:
-            data = self.ghJSON(["api", "user/orgs", "--paginate"])
-        except Exception as e:
-            logging.warning(f"Could not fetch user organizations: {e}")
-            return []
-        if not isinstance(data, list):
-            return []
-        return [org["login"] for org in data if isinstance(org, dict) and "login" in org]
-
-    def repoExists(self, nameWithOwner):
-        """Read-only check: does a repo at {owner}/{name} exist and is it visible to us?
-
-        Used as the pre-create / pre-publish collision preflight.  Creates nothing.  Returns
-        True if the GET succeeds, False on a 404 (or any error, treated as 'available')."""
-        try:
-            self.gh(["api", f"/repos/{nameWithOwner}", "--silent"])
-            return True
-        except RuntimeError:
-            return False
-
-    def setRepoVisibility(self, nameWithOwner, public=True):
-        """Flip a repository's visibility via the REST API.
-
-        Using the API avoids `gh repo edit --visibility`'s required
-        --accept-visibility-change-consequences confirmation flag."""
-        privateValue = "false" if public else "true"
-        self.gh(["api", "--method", "PATCH", f"/repos/{nameWithOwner}",
-                 "-F", f"private={privateValue}"])
-
-    def addMorphoTopics(self, nameWithOwner, speciesTopicString):
-        """Publish topic transition (last step of go-live): add the discoverability topic(s)
-        and REMOVE the `morphodepot-staging` marker, so the repo becomes findable by RepoClerk
-        and the extension and simultaneously leaves the unpublished list.  The species topic is
-        skipped when unknown (e.g. recovered repos with no species)."""
-        command = f"repo edit {nameWithOwner} --add-topic morphodepot --remove-topic {self.stagingTopic}"
-        if speciesTopicString:
-            command = (f"repo edit {nameWithOwner} --add-topic morphodepot "
-                       f"--add-topic md-{speciesTopicString} --remove-topic {self.stagingTopic}")
-        self.gh(command)
-
-    def issueList(self):
-        me = self.whoami()
-        repoData = self.ghTopicData()
-        issueList = []
-        for repo in repoData:
-            for issue in repo['issues']['nodes']:
-                assignees = [node['login'] for node in issue['assignees']['nodes']]
-                if me in assignees:
-                    repoName = repo['nameWithOwner'].split("/")[1]
-                    issueList.append({'number': issue['number'],
-                                      'title': issue['title'],
-                                      'repository': { 'name': repoName, 'nameWithOwner': repo['nameWithOwner']}})
-        return issueList
-
-    def administratedRepoList(self):
-        # Releases are ARCHIVAL-ONLY (org-design Sec.9.6): only org repos the user curates can be
-        # released, so the Release tab lists only those.  Personal short-term repos are excluded —
-        # they cannot be released, and showing them is misleading.  A member's archival repo is owned
-        # by the MorphoDepot org (owner != me); the journaled CURATOR (RepoClerk schema v3) is the
-        # authoritative "this is mine to release" signal.  (curator is None on pre-v3 journals, so an
-        # org repo lights up once RepoClerk re-drains with the field.)
-        me = self.whoami()
-        returnRepos = []
-        for repo in self.morphoRepos():
-            ownerLogin = repo['owner']['login']
-            isArchival = ownerLogin != me           # archival repos live in the org, not on my account
-            if isArchival and repo.get('curator') == me:
-                repo['nameWithOwner'] = f"{ownerLogin}/{repo['name']}"
-                returnRepos.append(repo)
-        return returnRepos
-
-    def prList(self, role="segmenter"):
-        """
-        Fetch a list of open pull requests for the user, either as 'segmenter' or reviewer.
-        Returns PRs, their associated issue titles, and repository topics.
-        """
-        me = self.whoami()
-        repoData = self.ghTopicData()
-        prList = []
-        for repo in repoData:
-            for pr in repo['pullRequests']['nodes']:
-                prAuthor = (pr.get('author') or {}).get('login')  # None for a deleted/ghost account
-                if role == "segmenter":
-                    parties = [prAuthor] if prAuthor else []
-                elif role == "reviewer":
-                    parties = [issue['repository']['owner']['login'] for issue in pr['closingIssuesReferences']['nodes']]
-                    # In-org (archival) repos are owned by the MorphoDepot org, not the curator's
-                    # login, so owner-keying alone hides every in-org PR from its curator.  Add the
-                    # journaled curator — the same owner-vs-curator fix administratedRepoList() applies.
-                    if repo.get('curator'):
-                        parties.append(repo['curator'])
-                else:
-                    raise BaseException(f"Unknown role {role}")
-                issueTitles = [issue['title'] for issue in pr['closingIssuesReferences']['nodes']]
-                if me in parties:
-                    repoName = repo['nameWithOwner'].split("/")[1]
-                    prList.append({'number': pr['number'],
-                                      'title': pr['title'],
-                                      'issueTitles': issueTitles,
-                                      'isDraft': pr['isDraft'],
-                                      'author': {'login': prAuthor},
-                                      'repository': { 'name': repoName, 'nameWithOwner': repo['nameWithOwner']}})
-        return prList
-
-
-    def repositoryList(self):
-        # --limit 1000: gh defaults to 30, which would make loadIssue's fork-exists check
-        # (forkExists = name in repositoryList) miss a real fork for users with many repos.
-        repositories = self.ghJSON("repo list --json name --limit 1000") or []
-        repositoryList = [r['name'] for r in repositories]
-        return repositoryList
-
-    def _ensureUpstream(self, canonicalNameWithOwner):
-        """Point the 'upstream' remote at the canonical repo (not the fork) — set explicitly rather
-        than inferred from origin, so reviewer/release actions always resolve the right base repo."""
-        url = f"https://github.com/{canonicalNameWithOwner}.git"
-        if "upstream" in [remote.name for remote in self.localRepo.remotes]:
-            self.localRepo.git.remote("set-url", "upstream", url)
-        else:
-            self.localRepo.create_remote("upstream", url)
-
-    def _userForkOf(self, sourceRepository):
-        """The active user's fork of sourceRepository as 'owner/name', or None.  Identifies the fork
-        by its PARENT (GitHub's fork list), so a suffixed fork (e.g. name-1, created when the user
-        already owned an unrelated repo of that name) is found, and an unrelated same-named repo is
-        never mistaken for the fork.  (R-1/R-2)"""
-        me = self.whoami()
-        name = sourceRepository.split("/")[1]
-        try:  # fast path: the usual case is a same-named fork at {me}/{name}
-            info = self.ghJSON(["repo", "view", f"{me}/{name}", "--json", "parent"])
-        except Exception:
-            info = None
-        parent = (info or {}).get("parent") or {}
-        parentLogin = (parent.get("owner") or {}).get("login")
-        # `gh repo view --json parent` gives parent.{name, owner.login} (no nameWithOwner), so build it.
-        if parentLogin and parent.get("name") and f"{parentLogin}/{parent['name']}" == sourceRepository:
-            return f"{me}/{name}"
-        # slow path: a differently-named (suffixed) fork — ask the source's fork list directly.
-        # Guarded like the fast path: on a gh/network error return None so loadIssue falls through to
-        # fork creation rather than aborting.
-        try:
-            out = self.gh(["api", f"repos/{sourceRepository}/forks", "--paginate",
-                           "--jq", f'.[] | select(.owner.login=="{me}") | .full_name']) or ""
-        except Exception as e:
-            logging.warning(f"Could not list forks of {sourceRepository}: {e}")
-            return None
-        forks = [line.strip() for line in out.splitlines() if line.strip()]
-        return forks[0] if forks else None
-
-    def loadIssue(self, issue, repoDirectory):
-        self.currentIssue = issue
-        self.progressMethod(f"Loading issue {issue} into {repoDirectory}")
-        issueNumber = issue['number']
-        branchName=f"issue-{issueNumber}"
-        sourceRepository = issue['repository']['nameWithOwner']
-        repositoryName = issue['repository']['name']
-        localDirectory = os.path.join(repoDirectory, f"{repositoryName}-{branchName}")
-
-        self.cacheOldVersion(localDirectory)
-
-        # Resolve where to clone from.  The contributor works in their FORK of the canonical repo;
-        # the fork is identified by its parent (not its name), so an unrelated same-named repo, or a
-        # suffixed fork (name-1) from a prior name collision / a deleted+recreated source, is never
-        # mistaken for it.  The owner of the canonical repo clones it directly (no fork).  (R-1)
-        me = self.whoami()
-        if sourceRepository.split("/")[0] == me:
-            cloneTarget = sourceRepository
-            isFork = False
-        else:
-            cloneTarget = self._userForkOf(sourceRepository)
-            if not cloneTarget:
-                self.gh(f"repo fork {sourceRepository} --clone=false")
-                cloneTarget = self._userForkOf(sourceRepository) or f"{me}/{repositoryName}"
-            isFork = True
-        self.gh(["repo", "clone", cloneTarget, localDirectory])
-        self.localRepo = git.Repo(localDirectory)
-        self._ensureUpstream(sourceRepository)
-
-        # D2: keep a genuine fork's default branch current with upstream (GitHub forks do not
-        # auto-sync, so a pre-existing fork drifts behind every merge/release).  Best-effort
-        # server-side fast-forward (no --force, so it cannot clobber a diverged fork); a failure
-        # never aborts — D1 below cuts the new branch from upstream/main regardless.
-        if isFork and cloneTarget != sourceRepository:
-            try:
-                self.gh(["repo", "sync", cloneTarget, "--source", sourceRepository])
-            except Exception as e:
-                logging.warning(f"Could not sync fork {cloneTarget} from {sourceRepository}: {e}")
-
-        originBranches = self.localRepo.remotes.origin.fetch()
-        originBranchIDs = [ob.name for ob in originBranches]
-        originBranchID = f"origin/{branchName}"
-
-        # D1: fetch upstream so a new issue branch can be cut from the current upstream default
-        # branch (the latest published baseline), not the fork's possibly-stale main.
-        try:
-            self.localRepo.remote("upstream").fetch()
-        except Exception as e:
-            logging.warning(f"Could not fetch upstream: {e}")
-
-        localIssueBranch = None
-        for branch in self.localRepo.branches:
-            if branch.name == branchName:
-                localIssueBranch = branch
-                break
-
-        logging.debug("Making new branch")
-        if originBranchID in originBranchIDs:
-            logging.debug("Checking out existing from origin")
-            self.localRepo.git.execute(f"git checkout --track {originBranchID}".split())
-        else:
-            # D1: branch off upstream/main (latest published state).  Fall back to origin/main
-            # only if upstream/main is somehow unavailable, preserving the prior behavior.
-            base = "upstream/main"
-            try:
-                self.localRepo.git.rev_parse("--verify", base)
-            except Exception:
-                base = "origin/main"
-            logging.debug("Nothing in origin for %s; creating it from %s", branchName, base)
-            self.localRepo.git.checkout(base)
-            self.localRepo.git.branch(branchName)
-            self.localRepo.git.checkout(branchName)
-
-        self.loadFromLocalRepository()
-
-    def loadPR(self, pr, repoDirectory):
-        baseRepo = pr['repository']['nameWithOwner']
-        baseName = pr['repository']['name']
-        branchName = pr['title']
-        # Resolve the ACTUAL head repo + branch from GitHub — never construct {author}/{name}: a
-        # contributor's fork may be suffixed (name-1) after a name collision, and an in-repo-branch
-        # PR has head == base.  Fall back to the constructed name if the query is unavailable.  (R-2)
-        headNameWithOwner = None
-        number = pr.get('number')
-        if number is not None:
-            head = None
-            try:
-                head = self.ghJSON(["pr", "view", str(number), "--repo", baseRepo,
-                                    "--json", "headRepository,headRepositoryOwner,headRefName"])
-            except Exception as e:
-                logging.warning(f"Could not resolve PR #{number} head repo: {e}")
-            if head:
-                headOwner = (head.get("headRepositoryOwner") or {}).get("login")
-                headName = (head.get("headRepository") or {}).get("name")
-                if headOwner and headName:
-                    headNameWithOwner = f"{headOwner}/{headName}"
-                if head.get("headRefName"):
-                    branchName = head["headRefName"]
-        if not headNameWithOwner:
-            headNameWithOwner = f"{pr['author']['login']}/{baseName}"  # legacy fallback
-        localDirectory = os.path.join(repoDirectory, f"{baseName}-{branchName}")
-        self.progressMethod(f"Loading PR from {headNameWithOwner} into {localDirectory}")
-
-        self.cacheOldVersion(localDirectory)
-
-        self.gh(["repo", "clone", headNameWithOwner, localDirectory])
-        self.localRepo = git.Repo(localDirectory)
-        self._ensureUpstream(baseRepo)
-        self.localRepo.remotes.origin.fetch()
-        self.localRepo.git.checkout(branchName)
-
-        self.loadFromLocalRepository(configuration="reviewer")
-        return True
-
-    def loadRepoForRelease(self, repoData):
-        repoName = repoData['name']
-        repoNameWithOwner = repoData['nameWithOwner'] # this is owner/name
-        localDirectory = os.path.join(self.localRepositoryDirectory(), repoName)
-
-        self.cacheOldVersion(localDirectory)
-
-        # clone the main repo, not a fork
-        self.gh(f"repo clone {repoNameWithOwner} {localDirectory}")
-
-        self.localRepo = git.Repo(localDirectory)
-        self.localRepo.git.checkout("main")
-        self.loadFromLocalRepository(remoteName="origin", configuration="release")
-        return True
-
-    def loadRepoForPreview(self, repoNameWithOwner):
-        repoName = repoNameWithOwner.split('/')[1]
-        localDirectory = os.path.join(self.localRepositoryDirectory(), repoName)
-
-        self.cacheOldVersion(localDirectory)
-
-        self.gh(f"repo clone {repoNameWithOwner} {localDirectory}")
-
-        self.localRepo = git.Repo(localDirectory)
-        self.localRepo.git.checkout("main")
-        self.loadFromLocalRepository(remoteName="origin", configuration="preview")
-        return True
-
-    def loadFromLocalRepository(self, remoteName="upstream", configuration="segment"):
-        localDirectory = self.localRepo.working_dir
-        branchName = self.localRepo.active_branch.name
-        remoteNameWithOwner = self.nameWithOwner(remoteName)
-
-        self.progressMethod(f"Loading {branchName} into {localDirectory}")
-
-        self.colorTableNode = None
-        try:
-            colorPath = glob.glob(f"{localDirectory}/*.csv")[0]
-            self.colorTableNode = slicer.util.loadColorTable(colorPath)
-        except IndexError:
-            try:
-                colorPath = glob.glob(f"{localDirectory}/*.ctbl")[0]
-                self.colorTableNode = slicer.util.loadColorTable(colorPath)
-            except IndexError:
-                self.progressMethod("No color table found")
-
-        # TODO: move from single volume file to segmentation specification json
-        volumePath = os.path.join(localDirectory, "source_volume")
-        if not os.path.exists(volumePath):
-            volumePath = os.path.join(localDirectory, "master_volume") # for backwards compatibility
-        volumeRef = open(volumePath).read().strip()
-        # The source_volume pointer is "releases/download/v1/{originalName}.nrrd";
-        # remember the original name so the UI can display it.
-        self.sourceVolumeName = os.path.basename(volumeRef).rsplit('.nrrd', 1)[0]
-        volumeURL = self.resolveVolumeURL(volumeRef, remoteNameWithOwner)
-        cacheDirectory = os.path.join(self.localRepositoryDirectory(), "MorphoDepotCaches", "Volumes")
-        os.makedirs(cacheDirectory, exist_ok=True)
-        nrrdPath = os.path.join(cacheDirectory, f"{remoteNameWithOwner.replace('/', '-')}-volume.nrrd")
-        checksum = None
-        checksumFilePath = os.path.join(localDirectory, "source_volume_checksum")
-        if os.path.exists(checksumFilePath):
-            with open(checksumFilePath) as fp:
-                checksum = fp.read().strip()
-        if not os.path.exists(nrrdPath):
-            slicer.util.downloadFile(volumeURL, nrrdPath, checksum=checksum)
-        volumeNode = slicer.util.loadVolume(nrrdPath)
-        # Keep a handle to the source volume: it is the reference geometry used to hash a
-        # baseline segmentation's voxels (release "baseline unchanged" check).
-        self.sourceVolumeNode = volumeNode
-
-        # Load all segmentations
-        segmentationNodesByName = {}
-        for segmentationPath in glob.glob(f"{localDirectory}/*.seg.nrrd"):
-            name = os.path.split(segmentationPath)[1].split(".")[0]
-            segmentationNodesByName[name] = slicer.util.loadSegmentation(segmentationPath)
-        # Default for the New release "baseline segmentation" picker: prefer a segmentation
-        # named "baseline" if present, otherwise fall back to whatever loaded.
-        self.baselineSegmentationNode = (
-            segmentationNodesByName.get("baseline")
-            or (next(iter(segmentationNodesByName.values()), None))
-        )
-
-        if configuration in ("segment", "reviewer"):
-            for segmentationNode in segmentationNodesByName.values():
-                segmentationNode.GetDisplayNode().SetVisibility(False)
-
-            # Switch to Segment Editor module
-            pluginHandlerSingleton = slicer.qSlicerSubjectHierarchyPluginHandler.instance()
-            pluginHandlerSingleton.pluginByName("Default").switchToModule("SegmentEditor")
-            editorWidget = slicer.modules.segmenteditor.widgetRepresentation().self()
-
-            self.segmentationPath = os.path.join(localDirectory, branchName) + ".seg.nrrd"
-            if branchName in segmentationNodesByName.keys():
-                self.segmentationNode = segmentationNodesByName[branchName]
-                self.segmentationNode.GetDisplayNode().SetVisibility(True)
-            else:
-                self.segmentationNode = slicer.mrmlScene.AddNewNodeByClass("vtkMRMLSegmentationNode")
-                self.segmentationNode.CreateDefaultDisplayNodes()
-                self.segmentationNode.SetReferenceImageGeometryParameterFromVolumeNode(volumeNode)
-                self.segmentationNode.SetName(branchName)
-                if "baseline" in segmentationNodesByName.keys():
-                    baselineSegmentation = segmentationNodesByName["baseline"].GetSegmentation()
-                    newSegmentation = self.segmentationNode.GetSegmentation()
-                    for segmentID in baselineSegmentation.GetSegmentIDs():
-                        newSegmentation.CopySegmentFromSegmentation(baselineSegmentation, segmentID)
-
-            editorWidget.parameterSetNode.SetAndObserveSegmentationNode(self.segmentationNode)
-            editorWidget.parameterSetNode.SetAndObserveSourceVolumeNode(volumeNode)
-
-    def nameWithOwner(self, remote):
-        branchName = self.localRepo.active_branch.name
-        repo = self.localRepo.remote(name=remote)
-        repoURL = list(repo.urls)[0]
-        if repoURL.find("@") != -1:
-            # git ssh prototocol
-            repoURL = "/".join(repoURL.split(":"))
-            repoNameWithOwner = "/".join(repoURL.split("/")[-2:]).split(".")[0]
-        elif repoURL.startswith("https://"):
-            # https protocol
-            repoNameWithOwner = "/".join(repoURL.split("/")[-2:]).split(".")[0]
-        elif repoURL.startswith("git@"):
-            # git@github.com:owner/repo.git
-            repoNameWithOwner = repoURL.split(":")[1].replace(".git", "")
-        elif os.path.exists(repoURL):
-            # local path
-            # this case happens during repo creation before pushing to remote
-            return None
-        else:
-            # https protocol
-            repoNameWithOwner = "/".join(repoURL.split("/")[-2:]).split(".")[0]
-        return repoNameWithOwner
-
-    def issuePR(self, role="segmenter"):
-        """Find the issue for the issue currently being worked on or None if there isn't one"""
-        if role not in ("segmenter", "reviewer"):
-            raise ValueError(f"Invalid role {role}")
-        if not self.localRepo:
-            return None
-        branchName = self.localRepo.active_branch.name
-        try:
-            upstreamNameWithOwner = self.nameWithOwner("upstream")
-        except ValueError:
-            return None
-        issuePR = None
-        prs = self.prList(role=role)
-        for pr in prs:
-            prRepoNameWithOwner = pr['repository']['nameWithOwner']
-            if prRepoNameWithOwner == upstreamNameWithOwner and pr['title'] == branchName:
-                issuePR = pr
-        return issuePR
-
-    def _openPRForBranch(self, upstreamNameWithOwner, headOwner, branchName):
-        """Authoritative check for an already-open PR whose head is headOwner:branchName into the
-        upstream repo, queried DIRECTLY from GitHub (gh api) rather than the lagging RepoClerk
-        journal that prList()/issuePR() read.  Returns the PR dict or None (None also on any query
-        error, so the caller falls through to PR creation, which is itself guarded against
-        duplicates).  The journal lags minutes behind GitHub, so it cannot answer "does a PR exist
-        for this branch right now" — the moment that matters when deciding whether to open one."""
-        try:
-            prs = self.ghJSON(["api",
-                               f"repos/{upstreamNameWithOwner}/pulls?head={headOwner}:{branchName}&state=open"])
-        except Exception as e:
-            logging.warning(f"Could not check for an existing PR on {branchName}: {e}")
-            return None
-        return prs[0] if prs else None
-
-    def cacheOldVersion(self, directoryPath):
-        """If directoryPath exists, move it to an archive in the cache."""
-        if os.path.exists(directoryPath):
-            self.progressMethod(f"Archiving old version of {directoryPath}")
-            cacheDirectory = os.path.join(self.localRepositoryDirectory(), "MorphoDepotCaches", "OldRepositories")
-            os.makedirs(cacheDirectory, exist_ok=True)
-            timestamp = time.strftime("%Y%m%d-%H%M%S")
-            archiveName = f"{os.path.basename(directoryPath)}-{timestamp}"
-            archivePath = os.path.join(cacheDirectory, archiveName)
-            shutil.move(directoryPath, archivePath)
-
-    def commitAndPush(self, message):
-        """Create a PR if needed and push current segmentation
-        Mark the PR as a draft
-        """
-        if not self.segmentationNode:
-            return False
-        if not slicer.util.saveNode(self.segmentationNode, self.segmentationPath, properties={'useCompression': True}):
-            logging.error(f"Segmentation save failed: path is {self.segmentationPath}")
-            return False
-        self.localRepo.index.add([self.segmentationPath])
-        self.localRepo.index.commit(message)
-
-        branchName = self.localRepo.active_branch.name
-        remote = self.localRepo.remote(name="origin")
-
-        # rebase branch if it exists in case other changes have been made (e.g. on another machine)
-        branchNames = [branch.name.split("/")[1] for branch in self.localRepo.remotes['origin'].refs]
-        if branchName in branchNames:
-            pullResult = self.localRepo.git.pull(f"--rebase", "origin", branchName)
-            self.progressMethod(pullResult)
-
-        # Workaround for missing origin.push().raise_if_error() in 3.1.14
-        # (see https://github.com/gitpython-developers/GitPython/issues/621):
-        # https://github.com/gitpython-developers/GitPython/issues/621
-        pushInfoList = remote.push(branchName)
-        for pi in pushInfoList:
-            for flag in [pi.REJECTED, pi.REMOTE_REJECTED, pi.REMOTE_FAILURE, pi.ERROR]:
-                if pi.flags & flag:
-                    self.progressMethod(f"Push failed with {flag}")
-                    return False
-
-        # Create a PR if one does not already exist.  Check AUTHORITATIVELY (direct gh), NOT via the
-        # RepoClerk journal that issuePR()/prList() read: the journal lags minutes behind GitHub, so
-        # right after the first push it still reports "no PR" and the old `if not self.issuePR()`
-        # guard would try to open a SECOND PR for the same head->base.  gh rejects that ("a pull
-        # request ... already exists"), which surfaced as a spurious "Failed to commit and push"
-        # even though the push succeeded -- for repo owners and outside segmenters alike.
-        upstreamNameWithOwner = self.nameWithOwner("upstream")
-        originNameWithOwner = self.nameWithOwner("origin")
-        originOwner = originNameWithOwner.split("/")[0]
-        if not self._openPRForBranch(upstreamNameWithOwner, originOwner, branchName):
-            issueNumber = branchName.split("-")[1]
-            prBody = f"Fixes #{issueNumber}"
-            if self.currentIssue and 'author' in self.currentIssue and 'login' in self.currentIssue['author']:
-                authorLogin = self.currentIssue['author']['login']
-                prBody = f"Started work on this issue for @{authorLogin}. {prBody}"
-            commandList = f"""
-                pr create
-                --draft
-                --repo {upstreamNameWithOwner}
-                --base main
-                --title {branchName}
-                --head {originOwner}:{branchName}
-            """.replace("\n"," ").split()
-            commandList += ["--body", prBody]
-            try:
-                self.gh(commandList)
-            except RuntimeError as e:
-                # Backstop for the race between the check above and now: gh refuses a duplicate
-                # head->base PR.  The push already updated that PR, so "already exists" is success;
-                # re-raise anything else.
-                if "already exists" not in str(e):
-                    raise
-                self.progressMethod("A pull request for this issue already exists; your push updated it.")
-            try:
-                self.notifyRepoClerk(upstreamNameWithOwner)
-            except Exception as e:
-                logging.warning(f"Could not notify RepoClerk: {e}")
-        return True
-
-    def requestReview(self):
-        upstreamNameWithOwner = self.nameWithOwner("upstream")
-        pr = self.issuePR(role="segmenter")
-        if not pr:
-            # The RepoClerk journal that issuePR() reads lags, so a just-opened PR may not appear
-            # there yet -- fall back to an authoritative direct lookup before giving up (same root
-            # cause as the commitAndPush duplicate-PR bug).
-            branchName = self.localRepo.active_branch.name
-            originOwner = self.nameWithOwner("origin").split("/")[0]
-            pr = self._openPRForBranch(upstreamNameWithOwner, originOwner, branchName)
-        if not pr:
-            logging.error("No pull request found for the current issue branch.")
-            return
-
-        self.gh(f"""
-            pr ready {pr['number']}
-                --repo {upstreamNameWithOwner}
-            """)
-        try:
-            self.notifyRepoClerk(upstreamNameWithOwner)
-        except Exception as e:
-            logging.warning(f"Could not notify RepoClerk: {e}")
-
-    def requestChanges(self, message=""):
-        pr = self.issuePR(role="reviewer")
-        if not pr:
-            branch = self.localRepo.active_branch.name if self.localRepo else "this branch"
-            raise RuntimeError(
-                f"No open pull request found for '{branch}'. It may already be merged or closed "
-                f"(the Review list can lag a minute behind GitHub). Click 'Refresh Github' to update it.")
-        upstreamNameWithOwner = self.nameWithOwner("upstream")
-        # GitHub forbids submitting a review (--request-changes) on your OWN pull request, exactly like
-        # --approve (see approvePR).  When the reviewer is also the PR author (their own contribution,
-        # or testing), post the feedback as a plain comment instead; the PR is still set back to draft
-        # below so the contributor knows to revise.
-        me = self.whoami()
-        selfAuthored = (pr.get("author") or {}).get("login") == me
-        if selfAuthored:
-            # Always post a comment -- an empty message would otherwise set the PR back to draft
-            # (below) with zero feedback to the contributor.
-            body = message if message != "" else "Changes requested (no additional comment)."
-            self.gh(["pr", "comment", str(pr['number']), "--repo", upstreamNameWithOwner,
-                     "--body", body])
-        else:
-            commandList = f"""
-                pr review {pr['number']}
-                    --request-changes
-                    --repo {upstreamNameWithOwner}
-            """.replace("\n"," ").split()
-            if message != "":
-                commandList += ["--body", message]
-            self.gh(commandList)
-        self.gh(f"""
-            pr ready {pr['number']}
-                --undo
-                --repo {upstreamNameWithOwner}
-            """)
-        try:
-            self.notifyRepoClerk(upstreamNameWithOwner)
-        except Exception as e:
-            logging.warning(f"Could not notify RepoClerk: {e}")
-
-    def approvePR(self, message=""):
-        pr = self.issuePR(role="reviewer")
-        if not pr:
-            branch = self.localRepo.active_branch.name if self.localRepo else "this branch"
-            raise RuntimeError(
-                f"No open pull request found for '{branch}'. It may already be merged or closed "
-                f"(the Review list can lag a minute behind GitHub). Click 'Refresh Github' to update it.")
-        upstreamNameWithOwner = self.nameWithOwner("upstream")
-        # GitHub forbids approving your OWN pull request (addPullRequestReview fails with
-        # "Can not approve your own pull request"), regardless of repo role — an "approve" review
-        # is by definition independent sign-off.  When the curator/reviewer is also the PR author
-        # (their own baseline contribution, or testing), skip the approval review and just merge:
-        # merging your own PR IS allowed for write/admin, only the review verdict is restricted.
-        me = self.whoami()
-        selfAuthored = (pr.get("author") or {}).get("login") == me
-        if not selfAuthored:
-            commandList = f"""
-                pr review {pr['number']}
-                    --approve
-                    --repo {upstreamNameWithOwner}
-            """.replace("\n"," ").split()
-            if message != "":
-                commandList += ["--body", message]
-            self.gh(commandList)
-        commandList = f"""
-            pr merge {pr['number']}
-                --repo {upstreamNameWithOwner}
-                --squash
-        """.replace("\n"," ").split()
-        commandList += ["--body", message if (selfAuthored and message) else "Merging and closing"]
-        self.gh(commandList)
-        try:
-            self.notifyRepoClerk(upstreamNameWithOwner)
-        except Exception as e:
-            logging.warning(f"Could not notify RepoClerk: {e}")
-
-    def getReleases(self):
-        """Get list of releases for the current repository (latest first)."""
-        if not self.localRepo:
-            return None
-        originNameWithOwner = self.nameWithOwner("origin")
-        return self.ghJSON(f"release list --repo {originNameWithOwner} --json name,tagName,publishedAt")
-
-    def closedIssuesSinceLastRelease(self, nameWithOwner):
-        """Return a list of {number,title} for issues closed since the last published release.
-        If there is no prior release, returns all closed issues for the repo.  Pre-release
-        ANNOUNCEMENT issues are excluded: they carry the invisible release-announce marker and are
-        not contributions -- retiring an announcement closes it, which would otherwise make it show
-        up as a 'change in this release' in the next cycle (the bug this filters out)."""
-        releases = self.ghJSON(f"release list --repo {nameWithOwner} --json tagName,publishedAt") or []
-        sinceDate = releases[0].get('publishedAt') if releases else None
-        if sinceDate:
-            cmd = ["issue", "list", "--repo", nameWithOwner,
-                   "--json", "number,title,body",
-                   "--search", f"is:issue is:closed closed:>{sinceDate}"]
-        else:
-            cmd = ["issue", "list", "--repo", nameWithOwner, "--state", "closed",
-                   "--json", "number,title,body"]
-        issues = self.ghJSON(cmd) or []
-        return [{"number": i["number"], "title": i["title"]} for i in issues
-                if self.releaseAnnounceMarkerName not in (i.get("body") or "")]
-
-    def nextReleaseTag(self):
-        """Compute the next vN tag based on existing releases. Returns None if no repo loaded."""
-        if not self.localRepo:
-            return None
-        releases = self.getReleases() or []
-        nextVersion = 1
-        tagNames = [r['tagName'] for r in releases if r['tagName'].startswith('v')]
-        versions = [int(t[1:]) for t in tagNames if t[1:].isdigit()]
-        if versions:
-            nextVersion = max(versions) + 1
-        return f"v{nextVersion}"
-
-    def previousReleaseTag(self):
-        """Latest existing release tag, or None if no releases yet."""
-        if not self.localRepo:
-            return None
-        releases = self.getReleases() or []
-        tagNames = [r['tagName'] for r in releases if r['tagName'].startswith('v')]
-        versions = [int(t[1:]) for t in tagNames if t[1:].isdigit()]
-        if not versions:
-            return None
-        return f"v{max(versions)}"
-
-    def existingScreenshotCount(self):
-        """Highest existing screenshot-N.png index in the loaded repo's screenshots/ directory."""
-        if not self.localRepo:
-            return 0
-        screenshotsDir = os.path.join(self.localRepo.working_dir, "screenshots")
-        if not os.path.exists(screenshotsDir):
-            return 0
-        highest = 0
-        for entry in os.listdir(screenshotsDir):
-            match = re.match(r"^screenshot-(\d+)\.png$", entry)
-            if match:
-                highest = max(highest, int(match.group(1)))
-        return highest
-
-    def releaseSnapshotPlan(self, newTag, baselineNode, colorTableNode, screenshots):
-        """Build a description of what prepareReleaseSnapshot will do, for the confirmation UI."""
-        if not self.localRepo:
-            return None
-        repoDir = self.localRepo.working_dir
-        previousTag = self.previousReleaseTag()
-        issueSegFiles = sorted(os.path.basename(p) for p in glob.glob(os.path.join(repoDir, "issue-*.seg.nrrd")))
-        startIndex = self.existingScreenshotCount() + 1
-        return {
-            'newTag': newTag,
-            'previousTag': previousTag,
-            'baselineName': baselineNode.GetName() if baselineNode else None,
-            'colorTableName': colorTableNode.GetName() if colorTableNode else None,
-            'newScreenshotNames': [f"screenshot-{startIndex + i}.png" for i in range(len(screenshots or []))],
-            'issueSegFiles': issueSegFiles,
-            'archivedReadme': f"README-{previousTag}.md" if previousTag else None,
-        }
-
-    def generateReleaseReadme(self, newTag, newScreenshotEntries):
-        """Build a new README.md body for the given release. Reuses metadata from
-        MorphoDepotAccession.json. Links each previous release to its tag's tree on
-        GitHub so the reader sees the repository at that release stage. Lists only
-        screenshots from this release flow (older screenshots stay in their archived READMEs)."""
-        repoDir = self.localRepo.working_dir
-        accessionPath = os.path.join(repoDir, "MorphoDepotAccession.json")
-        accession = {}
-        if os.path.exists(accessionPath):
-            try:
-                with open(accessionPath) as f:
-                    accession = json.load(f)
-            except Exception as e:
-                logging.warning(f"Could not parse MorphoDepotAccession.json: {e}")
-
-        def field(key, default=""):
-            v = accession.get(key)
-            if isinstance(v, list) and len(v) >= 2:
-                return v[1] or default
-            return v or default
-
-        species = field('species', "Unknown species")
-        modality = field('modality', "Unknown")
-        contrast = field('contrastEnhancement', "Unknown")
-        scanDimensions = accession.get('scanDimensions', "Unknown")
-        scanSpacing = accession.get('scanSpacing', "Unknown")
-
-        try:
-            originNameWithOwner = self.nameWithOwner("origin")
-        except Exception:
-            originNameWithOwner = None
-
-        lines = [f"# Release {newTag}", ""]
-
-        previousReadmes = sorted(
-            glob.glob(os.path.join(repoDir, "README-v*.md")),
-            key=lambda p: int(os.path.basename(p).replace("README-v", "").replace(".md", "")),
-        )
-        versions = []
-        for p in previousReadmes:
-            nStr = os.path.basename(p).replace("README-v", "").replace(".md", "")
-            if nStr.isdigit():
-                versions.append(int(nStr))
-        newTagN = int(newTag[1:]) if newTag.startswith('v') and newTag[1:].isdigit() else None
-        if versions:
-            lines.append("## Previous releases")
-            # Each previous version vN links to the pre-release-v(N+1) branch's README.md:
-            # that branch captured main right before v(N+1) was prepped, i.e. the README as
-            # it was while vN was the most recent release (and reflects any edits made to
-            # README.md between vN and v(N+1)).
-            for i, v in enumerate(versions):
-                nextN = versions[i + 1] if i + 1 < len(versions) else newTagN
-                if originNameWithOwner and nextN is not None:
-                    url = f"https://github.com/{originNameWithOwner}/blob/pre-release-v{nextN}/README.md"
-                    lines.append(f"- [v{v}]({url})")
-                else:
-                    lines.append(f"- [v{v}](README-v{v}.md)")
-            lines.append("")
-
-        lines.append("## MorphoDepot Repository")
-        lines.append("Repository for segmentation of a specimen scan. See [this JSON file](MorphoDepotAccession.json) for specimen details.")
-        lines.append(f"* Species: {species}")
-        lines.append(f"* Modality: {modality}")
-        lines.append(f"* Contrast: {contrast}")
-        lines.append(f"* Dimensions: {scanDimensions}")
-        lines.append(f"* Spacing (mm): {scanSpacing}")
-
-        if newScreenshotEntries:
-            lines.append("")
-            lines.append(f"## Screenshots for {newTag}")
-            for name, caption in newScreenshotEntries:
-                altText = caption or name
-                lines.append(f"![{altText}](screenshots/{name})")
-                if caption:
-                    lines.append(f"_{caption}_")
-
-        # Contributors section, generated from CONTRIBUTORS.json (written just before this during the
-        # release; org-design Sec.9.6). The DOI block is added separately by the App at mint time.
-        try:
-            import MorphoDepotContributors as MDC
-            contribPath = os.path.join(repoDir, "CONTRIBUTORS.json")
-            if os.path.exists(contribPath):
-                creditMarkdown = MDC.render_markdown(MDC.load(contribPath))
-                if creditMarkdown:
-                    lines.append("")
-                    lines.append(creditMarkdown.rstrip())
-        except Exception as e:
-            logging.warning(f"Could not render contributors section: {e}")
-
-        return "\n".join(lines) + "\n"
-
-    def prepareReleaseSnapshot(self, newTag, baselineNode, colorTableNode, screenshots, candidateBranch=None):
-        """Stage the working tree for a release tag: write baseline.seg.nrrd from the picked
-        segmentation, overwrite the repo's color table with the picked one, rotate README.md
-        to README-{previousTag}.md and generate a fresh README.md, append new screenshots
-        with sequential numbering (and update screenshots/captions.json), drop issue-*.seg.nrrd
-        from the working tree (still in git history), and commit.
-
-        If `candidateBranch` is given (org/gated, Option C), the release commit is force-pushed to
-        that branch and local main + working tree are reset back to the pre-release state -- main is
-        NOT modified (the App archives main and fast-forwards it to the candidate on approval).
-        Otherwise (personal) the commit is pushed straight to origin/main as before."""
-        if not self.localRepo:
-            return None
-        repoDir = self.localRepo.working_dir
-        previousTag = self.previousReleaseTag()
-        baseSha = self.localRepo.head.commit.hexsha  # pre-release main, to restore in candidate mode
-
-        # Baseline segmentation
-        baselinePath = os.path.join(repoDir, "baseline.seg.nrrd")
-        if not slicer.util.saveNode(baselineNode, baselinePath, properties={'useCompression': True}):
-            raise RuntimeError(f"Failed to save baseline segmentation to {baselinePath}")
-
-        # Color table — overwrite the existing repo color file (.csv preferred over .ctbl).
-        csvPaths = glob.glob(f"{repoDir}/*.csv")
-        ctblPaths = glob.glob(f"{repoDir}/*.ctbl")
-        if csvPaths:
-            colorTablePath = csvPaths[0]
-        elif ctblPaths:
-            colorTablePath = ctblPaths[0]
-        else:
-            colorTablePath = os.path.join(repoDir, f"{colorTableNode.GetName()}.csv")
-        if not slicer.util.saveNode(colorTableNode, colorTablePath):
-            raise RuntimeError(f"Failed to save color table to {colorTablePath}")
-
-        # New screenshots — continue sequential numbering from existing files.
-        newScreenshotEntries = []
-        if screenshots:
-            screenshotsDir = os.path.join(repoDir, "screenshots")
-            os.makedirs(screenshotsDir, exist_ok=True)
-            startIndex = self.existingScreenshotCount() + 1
-            captionsPath = os.path.join(screenshotsDir, "captions.json")
-            captions = {}
-            if os.path.exists(captionsPath):
-                try:
-                    with open(captionsPath) as f:
-                        captions = json.load(f)
-                except Exception:
-                    captions = {}
-            for i, ss in enumerate(screenshots):
-                name = f"screenshot-{startIndex + i}.png"
-                shutil.copy(ss['path'], os.path.join(screenshotsDir, name))
-                captions[name] = ss.get('caption', '')
-                newScreenshotEntries.append((name, ss.get('caption', '')))
-            with open(captionsPath, "w") as f:
-                json.dump(captions, f, indent=2)
-
-        # README rotation + fresh README for the new tag
-        readmePath = os.path.join(repoDir, "README.md")
-        if previousTag and os.path.exists(readmePath):
-            shutil.move(readmePath, os.path.join(repoDir, f"README-{previousTag}.md"))
-        with open(readmePath, "w") as f:
-            f.write(self.generateReleaseReadme(newTag, newScreenshotEntries))
-
-        # Drop per-issue segmentations from the working tree (kept in history)
-        for path in glob.glob(os.path.join(repoDir, "issue-*.seg.nrrd")):
-            os.remove(path)
-
-        # Stage everything (added, modified, deleted), commit.
-        self.localRepo.git.add("--all")
-        self.localRepo.index.commit(f"Prepare release {newTag}")
-        if candidateBranch:
-            # Option C: publish the release commit to the candidate branch and leave main UNTOUCHED.
-            # Always reset local main + working tree back to the pre-release state afterward (even if
-            # the push fails), so an unapproved or re-submitted release never rewrites main.  The App
-            # archives main and fast-forwards it to this candidate on approval.
-            try:
-                self.localRepo.git.push("--force", "origin", f"HEAD:refs/heads/{candidateBranch}")
-            finally:
-                self.localRepo.git.reset("--hard", baseSha)
-        else:
-            self.localRepo.remote(name="origin").push("main")
-
-    def resetToReleaseBackup(self):
-        """Hard-reset main to the pre-release archive branch and discard untracked changes.
-        Local-only: does NOT undo anything that was already pushed to origin/main, and does
-        NOT delete the archive branch (it is intentionally kept as a permanent record)."""
-        backup = getattr(self, 'releaseBackupBranch', None)
-        if not (self.localRepo and backup):
-            return False
-        self.localRepo.git.reset("--hard", backup)
-        self.localRepo.git.clean("-fd")
-        self.releaseBackupBranch = None
-        return True
-
-    def discardReleaseBackup(self):
-        """Clear the in-memory reference to the archive branch after a successful release.
-        The branch itself is intentionally kept (locally and on origin) as a permanent
-        archive of the pre-release state."""
-        self.releaseBackupBranch = None
-
-    def createRelease(self, releaseNotes="", baselineSegmentationNode=None, colorTableNode=None, screenshots=None):
-        """Create a new release.
-
-        Org/archival repo (gated, Option C): the release commit is published to a
-        `release-candidate-{tag}` branch and **main is left untouched**.  On approval the App archives
-        the current main as `pre-release-{tag}` (preserving the per-issue contributions), fast-forwards
-        main to the candidate, cuts the tag, and mints the DOI.  This keeps main/history intact through
-        an unapproved or re-submitted (request-changes) release.  Returns a pending marker dict.
-
-        Personal/short-term repo: no gate -- archive main, rewrite main directly, and create the tag.
-        Returns the new tag."""
-        if not self.localRepo:
-            return None
-        if baselineSegmentationNode is None or colorTableNode is None:
-            raise RuntimeError("A baseline segmentation and color table must both be selected.")
-        tag = self.nextReleaseTag()
-        if tag is None:
-            return None
-        originNameWithOwner = self.nameWithOwner("origin")
-        if releaseNotes == "":
-            releaseNotes = f"Version {tag} release."
-        originOwner = originNameWithOwner.split("/")[0]
-
-        if originOwner == self.morphoDepotOrg:
-            # Build the release commit on the candidate branch (main untouched), then request review.
-            candidate = f"release-candidate-{tag}"
-            self.prepareReleaseSnapshot(tag, baselineSegmentationNode, colorTableNode,
-                                        screenshots or [], candidateBranch=candidate)
-            repoName = originNameWithOwner.split("/")[1]
-            resp = self.controlPlaneRequest(
-                "repos/release", {"repo": repoName, "tag": tag, "notes": releaseNotes}) or {}
-            if resp.get("status") == "pending_review":
-                return {"pending": True, "tag": tag, "reviewSentTo": resp.get("review_sent_to")}
-            return tag
-
-        # Personal/short-term repo: archive main, rewrite main directly, then create the tag.
-        backupName = f"pre-release-{tag}"
-        if backupName in [head.name for head in self.localRepo.heads]:
-            self.localRepo.git.branch("-D", backupName)
-        self.localRepo.git.branch(backupName)
-        self.releaseBackupBranch = backupName
-        self.localRepo.remote(name="origin").push(backupName)
-        self.prepareReleaseSnapshot(tag, baselineSegmentationNode, colorTableNode, screenshots or [])
-        commandList = ["release", "create", tag, "--repo", originNameWithOwner]
-        commandList += ["--notes", releaseNotes]
-        self.gh(commandList)
-        return tag
-
-    def openIssuesAndPRs(self, nameWithOwner):
-        """Return (issues, prs) lists of open items for the given repo, each with number and title."""
-        issues = self.ghJSON(f"issue list --repo {nameWithOwner} --state open --json number,title")
-        prs = self.ghJSON(f"pr list --repo {nameWithOwner} --state open --json number,title")
-        return issues, prs
-
-    def announceUpcomingRelease(self, nameWithOwner, deadlineISO, message):
-        """Announce an upcoming release.  Posts a comment on every open issue and PR (so people
-        who watch notifications are pinged) AND creates a dedicated, pinned, `release-pending`
-        announcement issue (so people who only visit the repo see it).  The announcement issue
-        body carries an invisible marker encoding the target tag and deadline, which is the
-        repo-state signal `findReleaseAnnouncement` later reads.  Substitutes {deadline} in the
-        message body. Returns (issueCount, prCount)."""
-        issues, prs = self.openIssuesAndPRs(nameWithOwner)
-        body = message.replace("{deadline}", deadlineISO)
-        for issue in issues:
-            n = str(issue['number'])
-            self.progressMethod(f"Announcement on issue #{n}: {issue['title']}")
-            self.gh(["issue", "comment", n, "--repo", nameWithOwner, "--body", body])
-        for pr in prs:
-            n = str(pr['number'])
-            self.progressMethod(f"Announcement on PR #{n}: {pr['title']}")
-            self.gh(["pr", "comment", n, "--repo", nameWithOwner, "--body", body])
-        self._upsertReleaseAnnouncementIssue(nameWithOwner, deadlineISO, body)
-        return len(issues), len(prs)
-
-    def _releaseAnnounceMarker(self, tag, deadlineISO):
-        """Invisible HTML-comment marker embedded in the announcement issue body."""
-        return f"<!-- {self.releaseAnnounceMarkerName} tag={tag} deadline={deadlineISO} -->"
-
-    def _ensureReleasePendingLabel(self, nameWithOwner):
-        """Create (or update) the release-pending label; idempotent via --force."""
-        try:
-            self.gh(["label", "create", self.releasePendingLabel, "--repo", nameWithOwner,
-                     "--color", "FBCA04",
-                     "--description", "An upcoming release has been announced but not yet cut",
-                     "--force"])
-        except Exception as e:
-            logging.warning(f"Could not ensure '{self.releasePendingLabel}' label: {e}")
-
-    def _upsertReleaseAnnouncementIssue(self, nameWithOwner, deadlineISO, body):
-        """Update the existing release-pending announcement issue in place (new title/body/
-        deadline), or create+pin one if none exists. Editing in place — rather than close +
-        recreate — keeps a single, stable announcement issue. Best-effort: a failure here must
-        not abort the announcement comments above."""
-        tag = self.nextReleaseTag() or ""
-        self._ensureReleasePendingLabel(nameWithOwner)
-        marker = self._releaseAnnounceMarker(tag, deadlineISO)
-        title = f"Upcoming release {tag} - finish by {deadlineISO}".strip()
-        fullBody = f"{body}\n\n{marker}"
-        existing = self.findReleaseAnnouncement(nameWithOwner)
-        if existing:
-            n = str(existing["number"])
-            try:
-                self.gh(["issue", "edit", n, "--repo", nameWithOwner, "--title", title, "--body", fullBody])
-            except Exception as e:
-                logging.warning(f"Could not update announcement issue #{n}: {e}")
-            return n
-        try:
-            url = self.gh(["issue", "create", "--repo", nameWithOwner,
-                           "--title", title, "--body", fullBody,
-                           "--label", self.releasePendingLabel])
-        except Exception as e:
-            logging.warning(f"Could not create announcement issue: {e}")
-            return None
-        number = url.strip().rstrip("/").split("/")[-1]
-        try:
-            self.gh(["issue", "pin", number, "--repo", nameWithOwner])
-        except Exception as e:
-            logging.warning(f"Could not pin announcement issue #{number}: {e}")
-        return number
-
-    def listReleaseAnnouncements(self, nameWithOwner):
-        """All open release-pending announcement issues as [{'number','deadline'}], oldest first.
-        Returns **None** if the check itself FAILED (e.g. transient gh/network error) vs **[]**
-        when there genuinely are none — so callers can tell "couldn't check" from "none exist"
-        and avoid falsely clearing UI state.  Normally 0 or 1, but a repo can carry duplicates
-        from before the dedup guard."""
-        try:
-            items = self.ghJSON(["issue", "list", "--repo", nameWithOwner, "--state", "open",
-                                 "--label", self.releasePendingLabel, "--json", "number,body"])
-        except Exception as e:
-            logging.warning(f"Could not list release announcements: {e}")
-            return None
-        found = []
-        for item in items or []:
-            body = item.get("body", "") or ""
-            match = re.search(re.escape(self.releaseAnnounceMarkerName) + r"\b([^>]*)", body)
-            if match:
-                deadlineMatch = re.search(r"deadline=(\S+)", match.group(1))
-                found.append({"number": item["number"],
-                              "deadline": deadlineMatch.group(1) if deadlineMatch else None})
-        return sorted(found, key=lambda a: a["number"])
-
-    def findReleaseAnnouncement(self, nameWithOwner):
-        """The open release-pending announcement {'number','deadline'}, or None (the first, if
-        several exist).  Reads repo state only."""
-        announcements = self.listReleaseAnnouncements(nameWithOwner)
-        return announcements[0] if announcements else None
-
-    def clearReleaseAnnouncement(self, nameWithOwner, tag=None, message=None):
-        """Retire EVERY open pre-release announcement: unpin, remove the release-pending label,
-        comment, and close each.  Used both after a release (default message) and when an
-        announcement is superseded by a new one (caller passes `message`).  Clears ALL of them
-        (not just the oldest) so a repo carrying duplicates is fully cleaned up.  Best-effort
-        and idempotent."""
-        if message is None:
-            message = (f"Release {tag} has been published; closing this announcement." if tag
-                       else "The release has been published; closing this announcement.")
-        for announcement in (self.listReleaseAnnouncements(nameWithOwner) or []):
-            n = str(announcement["number"])
-            for command in (
-                ["issue", "unpin", n, "--repo", nameWithOwner],
-                ["issue", "edit", n, "--repo", nameWithOwner, "--remove-label", self.releasePendingLabel],
-                ["issue", "comment", n, "--repo", nameWithOwner, "--body", message],
-                ["issue", "close", n, "--repo", nameWithOwner],
-            ):
-                try:
-                    self.gh(command)
-                except Exception as e:
-                    logging.warning(f"Announcement cleanup step failed (#{n} {command[1]}): {e}")
-
-    def closeOpenItemsForRelease(self, nameWithOwner, version, message=None):
-        """Comment on and close every open issue and PR. PRs are closed without merging.
-        Returns (issueCount, prCount)."""
-        if message is None:
-            message = (
-                f"Release {version} has been published. This will be closed.\n"
-                f"Open a new issue or PR on the updated baseline to continue contributing."
-            )
-        issues, prs = self.openIssuesAndPRs(nameWithOwner)
-        for issue in issues:
-            n = str(issue['number'])
-            self.progressMethod(f"Closing issue #{n}: {issue['title']}")
-            self.gh(["issue", "comment", n, "--repo", nameWithOwner, "--body", message])
-            self.gh(["issue", "close", n, "--repo", nameWithOwner])
-        for pr in prs:
-            n = str(pr['number'])
-            self.progressMethod(f"Closing PR #{n}: {pr['title']}")
-            self.gh(["pr", "comment", n, "--repo", nameWithOwner, "--body", message])
-            self.gh(["pr", "close", n, "--repo", nameWithOwner])
-        return len(issues), len(prs)
-
-    def _resolveSpeciesString(self, accessionData, fallbackSpecies=""):
-        """Determine the species string from accession data — via the iDigBio record when the
-        specimen is accessioned there, otherwise the directly-entered species field.  Shared by
-        create (_stageRepoFiles) and edit (saveStagedRepoEdits) so the README/topic stay
-        consistent.  Robust to an unavailable/empty iDigBio response (network error, bad
-        specimen id, service down): it never raises, and on the edit path `fallbackSpecies`
-        (the species already recorded for the repo) is used so a transient outage does not
-        blank a previously-resolved species."""
-        if accessionData.get('iDigBioAccessioned', ['', ''])[1] == "Yes":
-            idigbioURL = accessionData.get('iDigBioURL', ['', ''])[1]
-            specimenID = idigbioURL.split("/")[-1] if idigbioURL else ""
-            try:
-                import idigbio
-                api = idigbio.json()
-                idigbioData = api.view("records", specimenID) if specimenID else None
-                data = idigbioData.get('data') if isinstance(idigbioData, dict) else None
-                if data:
-                    if 'ala:species' in data:
-                        return data['ala:species']
-                    if 'dwc:scientificName' in data:
-                        return data['dwc:scientificName']
-                logging.warning(f"Could not find species for '{idigbioURL}' (response: {idigbioData})")
-            except Exception as e:
-                logging.warning(f"iDigBio species lookup failed for '{idigbioURL}': {e}")
-            # Lookup unavailable/empty: keep the already-recorded species (edit), then the
-            # directly-entered field, then a safe default — never crash, never blindly blank.
-            return fallbackSpecies or accessionData.get('species', ['', ''])[1] or "Unknown species"
-        return accessionData.get('species', ['', ''])[1]
-
-    def _writeLicense(self, repoDir, accessionData):
-        """Write LICENSE.txt for the chosen Creative Commons license."""
-        if accessionData["license"][1].startswith("CC BY-NC"):
-            licenseURL = "https://creativecommons.org/licenses/by-nc/4.0/legalcode.txt"
-        else:
-            licenseURL = "https://creativecommons.org/licenses/by/4.0/legalcode.txt"
-        response = requests.get(licenseURL)
-        with open(os.path.join(repoDir, "LICENSE.txt"), "w") as fp:
-            fp.write(response.content.decode('ascii', errors="ignore"))
-
-    def _renderReadme(self, accessionData, speciesString, screenshotItems=None):
-        """Build README.md text from accession data.  screenshotItems is an ordered list of
-        (filename, caption) for images under the screenshots/ directory."""
-        readme_content = f"""
-## MorphoDepot Repository
-Repository for segmentation of a specimen scan.  See [this JSON file](MorphoDepotAccession.json) for specimen details.
-* Species: {speciesString}
-* Modality: {accessionData['modality'][1]}
-* Contrast: {accessionData['contrastEnhancement'][1]}
-* Dimensions: {accessionData['scanDimensions']}
-* Spacing (mm): {accessionData['scanSpacing']}
-"""
-        if screenshotItems:
-            readme_content += "\n\n## Screenshots\n"
-            for filename, caption in screenshotItems:
-                readme_content += f"\n![{caption or filename}](screenshots/{filename})\n"
-                if caption:
-                    readme_content += f"_{caption}_\n"
-        return readme_content
-
-    def _readScreenshotCaptions(self, repoDir):
-        """Read an existing repo's screenshots/captions.json into an ordered list of
-        (filename, caption) so the README can be regenerated without the live screenshots."""
-        captionsPath = os.path.join(repoDir, "screenshots", "captions.json")
-        if not os.path.exists(captionsPath):
-            return []
-        try:
-            with open(captionsPath) as fp:
-                captions = json.load(fp)
-        except Exception:
-            return []
-        return [(name, captions[name]) for name in sorted(captions.keys())]
-
-    def _speciesFromReadme(self, repoDir):
-        """Extract the already-recorded species from a repo's committed README (the
-        '* Species: ...' line).  Used as the fallback when re-resolving species on edit so a
-        transient iDigBio outage cannot blank a previously-resolved species."""
-        readmePath = os.path.join(repoDir, "README.md")
-        if not os.path.exists(readmePath):
-            return ""
-        try:
-            with open(readmePath) as fp:
-                for line in fp:
-                    if line.strip().startswith("* Species:"):
-                        return line.split(":", 1)[1].strip()
-        except Exception:
-            pass
-        return ""
-
-    # Optional, opt-in at Create (gated on the `workflow` token scope): a GitHub Actions
-    # workflow that assigns each newly opened issue back to whoever opened it.  Uses the
-    # auto-provisioned GITHUB_TOKEN — no secrets to configure.  Verified to assign even
-    # non-collaborator authors on public repos.
-    autoAssignWorkflow = """name: Auto-assign issue to creator
-
-# When a new issue is opened, assign it to the person who opened it.
-# Uses the built-in GITHUB_TOKEN (auto-provisioned per run) — no secrets to configure.
-on:
-  issues:
-    types: [opened]
-
-jobs:
-  assign:
-    runs-on: ubuntu-latest
-    permissions:
-      issues: write
-    steps:
-      - name: Assign the new issue to its author
-        uses: actions/github-script@v7
-        with:
-          script: |
-            const author = context.payload.issue.user.login;
-            const res = await github.rest.issues.addAssignees({
-              owner: context.repo.owner,
-              repo: context.repo.repo,
-              issue_number: context.payload.issue.number,
-              assignees: [author],
-            });
-            const assigned = (res.data.assignees || []).map(a => a.login);
-            if (!assigned.includes(author)) {
-              core.warning(`'${author}' was not assigned (likely not an assignable user on this repo).`);
-            }
-"""
-
-    def _stageRepoFiles(self, repoDir, sourceVolume, colorTable, accessionData, sourceSegmentation=None, screenshots=None, useOrg=False, targetOwner=None, enableAutoAssign=False):
-        """Build the repository content on disk: save every file (including the CURATOR
-        file), `git init`, and make the initial commit.  No GitHub interaction beyond the
-        non-GitHub license/iDigBio lookups.  `enableAutoAssign` additionally writes a GitHub
-        Actions workflow that assigns new issues back to their creator (opt-in, scope-gated in
-        the UI).  Returns a build context dict consumed by provisionStagedRepo()."""
-
-        # The CURATOR is the person creating the repo and is responsible for reviewing its
-        # segmentation PRs.  It is always the creator, regardless of where the repo ends up.
-        curator = self.whoami()
-        repoName = os.path.basename(repoDir.rstrip(os.sep))
-
-        os.makedirs(repoDir)
-
-        # save data
-        repoFileNames = []
-        sourceFileName = sourceVolume.GetName()
-        sourceFilePath = os.path.join(repoDir, sourceFileName) + ".nrrd"
-        slicer.util.saveNode(sourceVolume, sourceFilePath, properties={'useCompression': True})
-
-        # Size cap depends on tier: members upload to S3 via multipart (10 GiB cap); non-members
-        # store the volume as a GitHub release asset (2 GiB limit). See docs/ObjectStorage-model.md.
-        sourceBytes = os.path.getsize(sourceFilePath)
-        sizeGB = sourceBytes / 2**30
-        if useOrg:
-            if sourceBytes > 10 * 2**30:
-                raise ValueError(
-                    f"This volume ({sizeGB:.1f} GB) exceeds the 10 GB limit; volumes that large are "
-                    "not currently supported. Crop or resample the volume.")
-        elif sourceBytes > 2 * 2**30:
-            if self.userIsOrgMember():
-                raise ValueError(
-                    f"This volume ({sizeGB:.1f} GB) exceeds the 2 GB limit for personal "
-                    "repositories. Create it under MorphoDepot instead — the org supports up to 10 GB.")
-            raise ValueError(
-                f"This volume is {sizeGB:.1f} GB. Personal repositories cap files at 2 GB (a GitHub "
-                "restriction). To publish volumes up to 10 GB, join the MorphoDepot organization and "
-                "create your repository there.")
-
-        # calculate and save checksum
-        checksum = slicer.util.computeChecksum('SHA256', sourceFilePath)
-        checksumFilePath = os.path.join(repoDir, "source_volume_checksum")
-        with open(checksumFilePath, "w") as fp:
-            fp.write(f"SHA256:{checksum}")
-
-        colorTableName = colorTable.GetName()
-        slicer.util.saveNode(colorTable, os.path.join(repoDir, colorTableName) + ".csv")
-        repoFileNames.append(f"{colorTableName}.csv")
-
-        # write accessionData file
-        accessionData['fileFormatVersion'] = MorphoDepotLogic.accessionFileFormatVersion
-        fp = open(os.path.join(repoDir, "MorphoDepotAccession.json"), "w")
-        fp.write(json.dumps(accessionData, indent=4))
-        fp.close()
-
-        # write license file
-        self._writeLicense(repoDir, accessionData)
-
-        speciesString = self._resolveSpeciesString(accessionData)
-        speciesTopicString = speciesString.lower().replace(" ", "-")
-
-        # write readme file
-        screenshotItems = None
-        if screenshots:
-            screenshotItems = [(f"screenshot-{i+1}.png", ss['caption']) for i, ss in enumerate(screenshots)]
-        readme_content = self._renderReadme(accessionData, speciesString, screenshotItems)
-        fp = open(os.path.join(repoDir, "README.md"), "w")
-        fp.write(readme_content)
-        fp.close()
-
-        # write CURATOR file: the GitHub handle of the person responsible for curating this
-        # repository and reviewing its segmentation PRs.  Always the creator, regardless of
-        # whether the repo eventually lives under a personal account or an organization.
-        with open(os.path.join(repoDir, "CURATOR"), "w") as fp:
-            fp.write(f"{curator}\n")
-
-        # create initial repo
-        repo = git.Repo.init(repoDir, initial_branch='main')
-
-        repoFileNames += [
-            "README.md",
-            "LICENSE.txt",
-            "MorphoDepotAccession.json",
-            "source_volume_checksum",
-            "CURATOR",
-        ]
-        if sourceSegmentation:
-            segmentationName = "baseline" # keyword used to detect segmentation to import when startin new issue
-            slicer.util.saveNode(sourceSegmentation, os.path.join(repoDir, segmentationName) + ".seg.nrrd")
-            repoFileNames.append(f"{segmentationName}.seg.nrrd")
-
-        if screenshots:
-            # Copy screenshots
-            screenshotsDir = os.path.join(repoDir, "screenshots")
-            os.makedirs(screenshotsDir, exist_ok=True)
-            for i, screenshotInfo in enumerate(screenshots):
-                newScreenshotName = f"screenshot-{i+1}.png"
-                newScreenshotPath = os.path.join(screenshotsDir, newScreenshotName)
-                shutil.copy(screenshotInfo['path'], newScreenshotPath)
-                repoFileNames.append(os.path.join("screenshots", newScreenshotName))
-
-            # Save captions to a file
-            captions = {f"screenshot-{i+1}.png": ss['caption'] for i, ss in enumerate(screenshots)}
-            captionsPath = os.path.join(screenshotsDir, "captions.json")
-            with open(captionsPath, "w") as f:
-                json.dump(captions, f, indent=2)
-            repoFileNames.append(os.path.join("screenshots", "captions.json"))
-        # Optional auto-assign workflow (opt-in at Create, only when the token has `workflow`
-        # scope — gated in the UI).  Committed like any other file so it survives the staged
-        # flow's history rewrites and ships with the published repo.
-        if enableAutoAssign:
-            workflowDir = os.path.join(repoDir, ".github", "workflows")
-            os.makedirs(workflowDir, exist_ok=True)
-            with open(os.path.join(workflowDir, "auto-assign.yml"), "w") as fp:
-                fp.write(MorphoDepotLogic.autoAssignWorkflow)
-            repoFileNames.append(os.path.join(".github", "workflows", "auto-assign.yml"))
-
-        repoFilePaths = [os.path.join(repoDir, fileName) for fileName in repoFileNames]
-        repo.index.add(repoFilePaths)
-        repo.index.commit("Initial commit")
-
-        # Best-effort duplicate-volume check while the staging progress is already up, so the
-        # network round-trip is hidden; surfaced passively in the Go-live status and at publish.
-        # Skipped for the developer self-test (targetOwner) — test volumes are not real data.
-        duplicateRepos = [] if targetOwner else self.duplicateVolumeRepos(checksum)
-
-        return {
-            "repoDir": repoDir,
-            "repo": repo,
-            "curator": curator,
-            "repoName": repoName,
-            "sourceFilePath": sourceFilePath,
-            "sourceFileName": sourceFileName,
-            "checksum": checksum,
-            "duplicateRepos": duplicateRepos,
-            "useOrg": useOrg,
-            "targetOwner": targetOwner,
-            "speciesTopicString": speciesTopicString,
-            "species": speciesString,
-            "committedFiles": repoFileNames,
-        }
-
-    def createAccessionRepo(self, sourceVolume, colorTable, accessionData, sourceSegmentation=None, screenshots=None, useOrg=None, targetOwner=None, enableAutoAssign=False):
-        """Stage a new accession repository: build it locally, then provision it.  `useOrg`
-        chooses the destination — True = born in the MorphoDepot org (members only; S3, 10 GB,
-        governed); False = the creator's personal account (GitHub release asset, 2 GB, fully
-        theirs).  When unspecified, defaults to the org for members.  `targetOwner` (set only by
-        the developer self-test) overrides routing: provision directly into that org via the
-        creator's own gh rights, non-member style (release asset, no App/S3).  Returns the staged
-        nameWithOwner."""
-        repoName = accessionData['githubRepoName'][1].split("/")[-1]
-        if useOrg is None:
-            useOrg = self.userIsOrgMember()
-        if targetOwner:
-            useOrg = False  # targetOwner forces the direct-to-org (non-member-style) path
-
-        # Fail fast on a GitHub name collision (the authoritative check) BEFORE building
-        # anything locally, so we never half-build for a name that is already taken.
-        curator = self.whoami()
-        # Collision-check the namespace the repo will actually land in.
-        if targetOwner:
-            target, where = f"{targetOwner}/{repoName}", f"the {targetOwner} organization"
-        elif useOrg:
-            target, where = f"{self.morphoDepotOrg}/{repoName}", f"the {self.morphoDepotOrg} organization"
-        else:
-            target, where = f"{curator}/{repoName}", f"your account ({curator})"
-        if self.repoExists(target):
-            raise ValueError(
-                f"A repository named '{repoName}' already exists in {where}. "
-                "If it is a repo you staged earlier but never published, reopen the MorphoDepot "
-                "module to resume or discard it, or delete it on GitHub; otherwise choose a "
-                "different name.")
-
-        # Local clones are disposable working copies.  Clear any stale leftover (e.g. from a
-        # previous interrupted attempt) so it never blocks a fresh build with the same name.
-        repoDir = os.path.join(self.localRepositoryDirectory(), repoName)
-        if os.path.exists(repoDir):
-            self.progressMethod(f"Removing stale local directory {repoDir}")
-            shutil.rmtree(repoDir, ignore_errors=True)
-
-        buildContext = self._stageRepoFiles(repoDir, sourceVolume, colorTable, accessionData,
-                                            sourceSegmentation, screenshots, useOrg=useOrg,
-                                            targetOwner=targetOwner, enableAutoAssign=enableAutoAssign)
-        return self.provisionStagedRepo(buildContext)
-
-    def _provisionStagedRepoInOrg(self, buildContext):
-        """Member tier: the App creates the repo IN-ORG (private, staged topic, {handle}-team
-        Write); we push the built content and upload the source volume to S3.  No personal
-        account, no transfer.  Returns the org nameWithOwner (MorphoDepot/<name>)."""
-        repoDir = buildContext["repoDir"]
-        repo = buildContext["repo"]
-        curator = buildContext["curator"]
-        repoName = buildContext["repoName"]
-        sourceFilePath = buildContext["sourceFilePath"]
-        sourceFileName = buildContext["sourceFileName"]
-        checksum = buildContext["checksum"]
-
-        self.progressMethod(f"Creating {self.morphoDepotOrg}/{repoName} (private) via the App...")
-        info = self.controlPlaneRequest("repos/create", {"name": repoName})
-        nameWithOwner = info["full_name"]
-        cloneURL = info["clone_url"]
-
-        # Push the locally built content to the empty in-org repo (member has Write via team).
-        self.localRepo = repo
-        try:
-            repo.create_remote("origin", cloneURL)
-        except Exception:
-            repo.remote(name="origin").set_url(cloneURL)
-        branchName = repo.active_branch.name
-        lastError = None
-        for delay in (0, 2, 4, 8, 16):
-            if delay:
-                self.progressMethod(f"Push retry in {delay}s...")
-                time.sleep(delay)
-            try:
-                repo.git.push("--set-upstream", "origin", branchName)
-                lastError = None
-                break
-            except Exception as pushError:
-                lastError = pushError
-        if lastError is not None:
-            raise RuntimeError(f"Push to {nameWithOwner} failed: {lastError}")
-
-        owner, name = nameWithOwner.split("/", 1)
-        try:
-            self.gh(f"api --method PUT /repos/{owner}/{name}/subscription --field subscribed=true --field ignored=false")
-        except Exception as e:
-            logging.warning(f"Could not subscribe to {nameWithOwner}: {e}")
-
-        # v1 release (version anchor, no asset)
-        self.gh(["release", "create", "--repo", nameWithOwner, "v1", "--notes", "Initial release"])
-
-        # Upload the source volume to S3 and record the absolute public URL.
-        publicURL = self.uploadSourceVolumeToObjectStore(
-            sourceFilePath, checksum, curator, repoName, f"{sourceFileName}.nrrd")
-        with open(os.path.join(repoDir, "source_volume"), "w") as fp:
-            fp.write(publicURL)
-        repo.index.add([f"{repoDir}/source_volume"])
-        repo.index.commit("Add source file url file")
-        repo.remote(name="origin").push()
-
-        self.stagingContext = {
-            "repoDir": repoDir,
-            "personalNameWithOwner": nameWithOwner,
-            "repoName": repoName,
-            "curator": curator,
-            "speciesTopicString": buildContext["speciesTopicString"],
-            "checksum": checksum,
-            "duplicateRepos": buildContext.get("duplicateRepos", []),
-            "isMember": True,
-        }
-        self.localRepo = None
-        if os.path.exists(repoDir):
-            shutil.rmtree(repoDir, ignore_errors=True)
-        return nameWithOwner
-
-    def provisionStagedRepo(self, buildContext):
-        """Provision the staged repo.  Members: born in-org via the App control plane
-        (_provisionStagedRepoInOrg).  Non-members: PRIVATE on the creator's personal account
-        (the path below).  The developer self-test (`targetOwner`) uses this same non-member path
-        but creates directly in the testing org via the creator's own gh rights.  Records staging
-        state and returns the nameWithOwner."""
-        if buildContext.get("useOrg"):
-            return self._provisionStagedRepoInOrg(buildContext)
-        repoDir = buildContext["repoDir"]
-        repo = buildContext["repo"]
-        curator = buildContext["curator"]
-        repoName = buildContext["repoName"]
-        sourceFilePath = buildContext["sourceFilePath"]
-        sourceFileName = buildContext["sourceFileName"]
-        checksum = buildContext["checksum"]
-
-        # The GitHub name collision was already checked in createAccessionRepo (before build).
-        # Owner is the creator's account by default, or an explicit targetOwner (testing org).
-        owner = buildContext.get("targetOwner") or curator
-        personalTarget = f"{owner}/{repoName}"
-
-        # Create PRIVATE: the staging state is invisible (private + no topic) until go-live.
-        try:
-            self.gh(f"repo create {personalTarget} --disable-wiki --private --source {repoDir} --push")
-        except RuntimeError as e:
-            # gh repo create --push can race with GitHub provisioning the new repo for
-            # git-over-HTTPS access; the create succeeds but the immediate push fails with
-            # "Repository not found". gh has already added the origin remote, so retry the
-            # push from the local clone with the branch specified (no upstream is set yet).
-            if "Repository not found" not in str(e):
-                raise
-            branchName = repo.active_branch.name
-            lastError = None
-            for delay in (2, 4, 8, 16):
-                self.progressMethod(f"Initial push raced with repo provisioning; retrying in {delay}s...")
-                time.sleep(delay)
-                try:
-                    repo.git.push("--set-upstream", "origin", branchName)
-                    lastError = None
-                    break
-                except Exception as pushError:
-                    lastError = pushError
-            if lastError is not None:
-                raise RuntimeError(f"Initial push retry failed after multiple attempts: {lastError}")
-
-        self.localRepo = repo
-        repoNameWithOwner = self.nameWithOwner("origin")
-
-        self.gh(f"repo edit {repoNameWithOwner} --enable-projects=false --enable-discussions=false")
-
-        # Tag the repo as staged-but-unpublished.  This topic is the durable, queryable record
-        # of staging state (no client-side marker); publish removes it.
-        self.gh(f"repo edit {repoNameWithOwner} --add-topic {self.stagingTopic}")
-
-        # subscribe to all notifications for the new repository
-        # gh repo watch was removed in newer gh CLI versions; use the API directly
-        owner, name = repoNameWithOwner.split("/", 1)
-        self.gh(f"api --method PUT /repos/{owner}/{name}/subscription --field subscribed=true --field ignored=false")
-
-        # Non-member tier: create the v1 release and upload the source volume AS A RELEASE ASSET
-        # (the volume lives on GitHub, capped at 2 GB). Members use S3 instead — see
-        # _provisionStagedRepoInOrg and docs/org-design.md §1.0.
-        commandList = ["release", "create", "--repo", repoNameWithOwner, "v1"]
-        commandList += ["--notes", "Initial release"]
-        self.gh(commandList)
-        self.gh(f"release upload --repo {repoNameWithOwner} v1 {sourceFilePath}#{sourceFileName}.nrrd")
-
-        # write source volume pointer: an owner-relative path resolved against the repo's current
-        # owner at read time (resolveVolumeURL).
-        fp = open(os.path.join(repoDir, "source_volume"), "w")
-        fp.write(f"releases/download/v1/{sourceFileName}.nrrd")
-        fp.close()
-
-        repo.index.add([f"{repoDir}/source_volume"])
-        repo.index.commit("Add source file url file")
-        repo.remote(name="origin").push()
-
-        self.stagingContext = {
-            "repoDir": repoDir,
-            "personalNameWithOwner": repoNameWithOwner,
-            "repoName": repoName,
-            "curator": curator,
-            "speciesTopicString": buildContext["speciesTopicString"],
-            "checksum": checksum,
-            "duplicateRepos": buildContext.get("duplicateRepos", []),
-        }
-
-        # The local working copy is disposable now that the repo is fully on GitHub — remove it
-        # (and the multi-GB source-volume file it holds) immediately.  Editing later re-clones
-        # via the reopen path.
-        self.localRepo = None
-        if os.path.exists(repoDir):
-            shutil.rmtree(repoDir, ignore_errors=True)
-        return repoNameWithOwner
-
-    def _publishStagedRepoInOrg(self, ctx):
-        """Member tier: ask the App to make the staged org repo public.  Going public is a governed
-        event (org-design 11.3): the App routes the request through the MD-reviewers review gate —
-        it emails a signed Approve link and does NOT flip the repo now.  Returns a pending-review
-        dict in that case (the repo stays private + staged until a reviewer approves, after which
-        the App makes it public on its own).  A non-gated immediate publish (backstop) returns the
-        nameWithOwner string, as before."""
-        repoName = ctx["repoName"]
-        species = ctx.get("speciesTopicString")
-        topics = ["morphodepot"] + ([f"md-{species}"] if species else [])
-        finalNameWithOwner = ctx["personalNameWithOwner"]
-        self.progressMethod(f"Submitting {finalNameWithOwner} for review...")
-        resp = self.controlPlaneRequest("repos/publish", {"repo": repoName, "topics": topics}) or {}
-        repoDir = ctx.get("repoDir")
-
-        if resp.get("status") == "pending_review":
-            # Review requested.  The repo is still private and still carries morphodepot-staging,
-            # so it remains in the unpublished list; drop only the transient local clone (reopen
-            # re-clones).  Do NOT notify RepoClerk (nothing public to crawl) or add a contact.
-            self.localRepo = None
-            if repoDir and os.path.exists(repoDir):
-                shutil.rmtree(repoDir, ignore_errors=True)
-            self.stagingContext = None
-            return {"pending": True, "nameWithOwner": finalNameWithOwner,
-                    "reviewSentTo": resp.get("review_sent_to")}
-
-        # Backstop: an immediate (non-gated) publish actually made it public — clean up as before.
-        self.ghTopicClearCache()
-        self.localRepo = None
-        if repoDir and os.path.exists(repoDir):
-            shutil.rmtree(repoDir, ignore_errors=True)
-        # The repo is now public + topic:morphodepot, so RepoClerk will crawl it — nudge it to
-        # journal the new repo promptly instead of waiting for the next scheduled sweep.
-        # Best-effort: the publish already succeeded, so a notify failure must not surface here.
-        try:
-            self.notifyRepoClerk(finalNameWithOwner)
-        except Exception as e:
-            logging.warning(f"Could not notify RepoClerk of publish: {e}")
-        self.stagingContext = None
-        return finalNameWithOwner
-
-    def publishStagedRepo(self):
-        """Publish the staged repo IN PLACE — it already lives at its final location (members:
-        in the MorphoDepot org via the App; non-members: their personal account; the developer
-        self-test: the testing org).  Members go through the App's one-way private->public
-        (_publishStagedRepoInOrg); everyone else flips public + adds discoverability topics
-        directly.  No transfer.  Returns the final nameWithOwner."""
-        ctx = getattr(self, "stagingContext", None)
-        if not ctx:
-            raise RuntimeError("No staged repository to publish.")
-        if ctx.get("isMember"):
-            return self._publishStagedRepoInOrg(ctx)
-        personal = ctx["personalNameWithOwner"]
-        speciesTopicString = ctx["speciesTopicString"]
-
-        # Flip public, then add the discoverability topic(s) and drop the staging topic — the
-        # moment the repo becomes discoverable and leaves the unpublished list.
-        self.setRepoVisibility(personal, public=True)
-        self.addMorphoTopics(personal, speciesTopicString)
-        self.ghTopicClearCache()
-
-        # The local working copy is no longer needed once published — remove it.
-        repoDir = ctx.get("repoDir")
-        self.localRepo = None
-        if repoDir and os.path.exists(repoDir):
-            shutil.rmtree(repoDir, ignore_errors=True)
-        # Now public + discoverable — nudge RepoClerk to journal it (best-effort, see above).
-        try:
-            self.notifyRepoClerk(personal)
-        except Exception as e:
-            logging.warning(f"Could not notify RepoClerk of publish: {e}")
-        self.stagingContext = None
-        return personal
-
-    def discardStagedRepo(self):
-        """Abandon the staged repo.  Deleting a repository needs the `delete_repo` token
-        scope, which we deliberately do not request — so instead of calling the API, we hand
-        the user off to the repo's GitHub Settings page (Danger Zone) to delete it from the
-        web, and clean up our own side: remove the local clone and the in-memory staging
-        state.  Returns the repo's settings URL for the caller to open.  Note: if the user
-        does not actually delete it, the repo keeps its `morphodepot-staging` topic and so
-        correctly stays in the unpublished list."""
-        ctx = getattr(self, "stagingContext", None)
-        if not ctx:
-            return None
-        personal = ctx["personalNameWithOwner"]
-        repoDir = ctx.get("repoDir")
-        if ctx.get("isMember"):
-            # Member tier: the App deletes the in-org repo AND its S3 object (only while private).
-            self.progressMethod(f"Discarding {personal} (deleting repo + volume)...")
-            self.controlPlaneRequest("repos/discard", {"repo": ctx["repoName"]})
-            self.localRepo = None
-            if repoDir and os.path.exists(repoDir):
-                shutil.rmtree(repoDir, ignore_errors=True)
-            self.stagingContext = None
-            return None    # already deleted — nothing for the UI to open
-        self.localRepo = None
-        if repoDir and os.path.exists(repoDir):
-            shutil.rmtree(repoDir, ignore_errors=True)
-        self.stagingContext = None
-        return f"https://github.com/{personal}/settings"
-
-    # --- Staged-repo recovery via the `morphodepot-staging` topic.  GitHub is the source of
-    # truth: a staged-but-unpublished repo is simply one carrying this topic.  No durable
-    # client state, so recovery works from any machine and survives a /tmp flush. ---
-
-    stagingTopic = "morphodepot-staging"
-
-    def listStagedRepos(self):
-        """Return the active user's repositories that are staged but not yet published,
-        identified by the `morphodepot-staging` topic.  Uses the repo LIST endpoint (topics
-        are reflected immediately, unlike the search index which lags for fresh repos), so it
-        is reliable right after staging and from any machine.  Returns marker-shaped dicts."""
-        me = self.whoami()
-        # Include org repos too — member-tier staged repos are born in MorphoDepot, owned by the
-        # org but accessible only to the member's {handle}-team.
-        try:
-            repos = self.ghJSON(["api", "/user/repos?affiliation=owner,organization_member&per_page=100", "--paginate"])
-        except Exception as e:
-            logging.warning(f"listStagedRepos: could not list repositories: {e}")
-            return []
-        staged = []
-        for repo in repos:
-            if not isinstance(repo, dict):
-                continue
-            owner = (repo.get("owner") or {}).get("login")
-            if owner not in (me, self.morphoDepotOrg):
-                continue
-            if self.stagingTopic not in (repo.get("topics") or []):
-                continue
-            name = repo.get("name")
-            nameWithOwner = repo.get("full_name") or f"{owner}/{name}"
-            staged.append({
-                "nameWithOwner": nameWithOwner,
-                "repoName": name,
-                "curator": me,
-                "repoDir": os.path.join(self.localRepositoryDirectory(), name),
-                "summary": nameWithOwner,
-            })
-        return staged
-
-    def _fetchAccessionData(self, nameWithOwner):
-        """Fetch and decode a repo's MorphoDepotAccession.json, or None if it has none.
-        Used to re-derive the species topic when resuming a repo from the unpublished list."""
-        try:
-            data = self.ghJSON(["api", f"/repos/{nameWithOwner}/contents/MorphoDepotAccession.json"])
-        except Exception:
-            return None
-        if not isinstance(data, dict) or "content" not in data:
-            return None
-        try:
-            import base64
-            decoded = base64.b64decode(data["content"]).decode("utf-8", errors="ignore")
-            return json.loads(decoded)
-        except Exception:
-            return None
-
-    def resumeStagedRepo(self, stagedRepo):
-        """Resume a repo chosen from the unpublished list: CLONE it fresh (so its accession
-        form can be pre-filled and edits applied), rebuild the in-memory staging context, and
-        return the loaded accession data.  Cloning is cheap — the source volume is a release
-        asset, not a git-tracked file — and gives a working tree for saveStagedRepoEdits()."""
-        nameWithOwner = stagedRepo.get("nameWithOwner")
-        repoName = stagedRepo.get("repoName")
-        repoDir = os.path.join(self.localRepositoryDirectory(), repoName)
-        if os.path.exists(repoDir):
-            shutil.rmtree(repoDir, ignore_errors=True)
-        self.gh(f"repo clone {nameWithOwner} {repoDir}")
-        self.localRepo = git.Repo(repoDir)
-
-        accession = {}
-        accessionPath = os.path.join(repoDir, "MorphoDepotAccession.json")
-        if os.path.exists(accessionPath):
-            try:
-                with open(accessionPath) as fp:
-                    accession = json.load(fp)
-            except Exception as e:
-                logging.warning(f"Could not read accession data for {nameWithOwner}: {e}")
-
-        species = ""
-        try:
-            species = (accession.get("species") or ["", ""])[1] or ""
-        except Exception:
-            species = ""
-        # Carry the committed checksum so the duplicate-volume warning works on reopen too
-        # (the clone persists here, unlike a fresh stage whose dir is deleted after provision).
-        checksum = None
-        checksumPath = os.path.join(repoDir, "source_volume_checksum")
-        if os.path.exists(checksumPath):
-            try:
-                with open(checksumPath) as fp:
-                    checksum = fp.read().strip()
-            except Exception as e:
-                logging.warning(f"Could not read source_volume_checksum for {nameWithOwner}: {e}")
-        self.stagingContext = {
-            "repoDir": repoDir,
-            "personalNameWithOwner": nameWithOwner,
-            "repoName": repoName,
-            "curator": stagedRepo.get("curator"),
-            "speciesTopicString": species.lower().replace(" ", "-") if species else "",
-            "checksum": checksum,
-            "duplicateRepos": self.duplicateVolumeRepos(checksum, exclude=nameWithOwner),
-            "isMember": nameWithOwner.split("/", 1)[0] == self.morphoDepotOrg,
-        }
-        return accession
-
-    def saveStagedRepoEdits(self, accessionData, colorTable=None, sourceSegmentation=None, screenshots=None):
-        """Apply edits to the currently-resumed staged repo (clone in stagingContext): rewrite
-        the metadata-derived files from `accessionData`, optionally replace the color table
-        and/or baseline segmentation, regenerate the screenshots from `screenshots` (None =
-        leave as-is), then — only if something actually changed — rewrite the repo's `main` as
-        a single clean commit and force-push it.  The source volume and its release asset are
-        NEVER touched (out of scope by design).  Recomputes the staging context's species topic
-        for the subsequent publish.  Returns True if a change was pushed, False if the repo was
-        already up to date."""
-        ctx = getattr(self, "stagingContext", None)
-        if not ctx:
-            raise RuntimeError("No staged repository is open to edit.")
-        repoDir = ctx.get("repoDir")
-        repo = self.localRepo
-        if not (repoDir and repo and os.path.exists(repoDir)):
-            raise RuntimeError("No local working copy is available to apply edits.")
-
-        # Capture the already-recorded species (from the committed README, before we rewrite
-        # it) so a transient iDigBio outage during re-resolution can't blank it.
-        priorSpecies = self._speciesFromReadme(repoDir)
-
-        # Regenerate the metadata-derived files from the edited data.
-        accessionData['fileFormatVersion'] = MorphoDepotLogic.accessionFileFormatVersion
-        with open(os.path.join(repoDir, "MorphoDepotAccession.json"), "w") as fp:
-            fp.write(json.dumps(accessionData, indent=4))
-
-        speciesString = self._resolveSpeciesString(accessionData, fallbackSpecies=priorSpecies)
-        ctx["speciesTopicString"] = speciesString.lower().replace(" ", "-") if speciesString else ""
-
-        self._writeLicense(repoDir, accessionData)
-
-        # Replace the color table only if a new one was supplied (else keep the committed CSV).
-        if colorTable is not None:
-            for existing in os.listdir(repoDir):
-                if existing.endswith(".csv"):
-                    os.remove(os.path.join(repoDir, existing))
-            slicer.util.saveNode(colorTable, os.path.join(repoDir, colorTable.GetName()) + ".csv")
-
-        # Replace the baseline segmentation only if a new one was supplied.
-        if sourceSegmentation is not None:
-            slicer.util.saveNode(sourceSegmentation, os.path.join(repoDir, "baseline.seg.nrrd"))
-
-        # Screenshots: when a set is supplied, fully regenerate screenshots/ + captions.json
-        # from it (add/remove/recaption); when None, leave the committed screenshots untouched.
-        screenshotsDir = os.path.join(repoDir, "screenshots")
-        if screenshots is not None:
-            if os.path.exists(screenshotsDir):
-                shutil.rmtree(screenshotsDir)
-            screenshotItems = []
-            if screenshots:
-                os.makedirs(screenshotsDir, exist_ok=True)
-                captions = {}
-                for i, ss in enumerate(screenshots):
-                    name = f"screenshot-{i+1}.png"
-                    shutil.copy(ss["path"], os.path.join(screenshotsDir, name))
-                    captions[name] = ss.get("caption", "")
-                    screenshotItems.append((name, ss.get("caption", "")))
-                with open(os.path.join(screenshotsDir, "captions.json"), "w") as f:
-                    json.dump(captions, f, indent=2)
-        else:
-            screenshotItems = self._readScreenshotCaptions(repoDir)
-
-        # Regenerate the README (screenshot section reflects the current screenshot set).
-        readme = self._renderReadme(accessionData, speciesString, screenshotItems)
-        with open(os.path.join(repoDir, "README.md"), "w") as fp:
-            fp.write(readme)
-
-        # Push only if the working tree actually changed.
-        repo.git.add("-A")
-        if not repo.git.diff("--cached", "--name-only").strip():
-            return False
-
-        # Reset history to a single clean commit ("as if created correctly") and force-push.
-        repo.git.checkout("--orphan", "_morphodepot_clean")
-        repo.git.add("-A")
-        repo.git.commit("-m", "MorphoDepot accession")
-        repo.git.branch("-M", "main")
-        repo.git.push("--force", "origin", "main")
-        return True
-
-    #
-    # Search
-    #
-
-    def refreshSearchCache(self):
-        """Gets accession data from all repositories via RepoClerk journals."""
-        # Reset up front so a failed refresh yields an empty cache (search then returns nothing)
-        # rather than silently serving stale results from a previous successful run.
-        self.repoDataByNameWithOwner = {}
-        clonePath = self.refreshRepoClerk()
-        if clonePath:
-            journals = self.repoClerkJournals(clonePath)
-            if journals:
-                for j in journals:
-                    try:
-                        owner, repo = j["nameWithOwner"].split("/", 1)
-                        key = f"{owner}^{repo}"
-                        repoData = dict(j.get("accession", {}))
-                        repoData["pushedAt"] = j.get("pushedAt", "")
-                        repoData["screenshotCount"] = j.get("screenshotCount", 0)
-                        repoData["screenshotCaptions"] = j.get("screenshotCaptions", [])
-                        repoData["volumeSize"] = j.get("volumeSize")
-                        self.repoDataByNameWithOwner[key] = repoData
-                        self.progressMethod(f"Loaded {key} from RepoClerk")
-                    except Exception as e:
-                        logging.warning(f"Could not process journal {j.get('nameWithOwner', '?')}: {e}")
-                self.progressMethod("Finished loading from RepoClerk")
-                return
-        logging.warning("RepoClerk journals unavailable — search cache not populated")
-
-
-    def search(self, criteria):
-        if self.repoDataByNameWithOwner == {}:
-            return {}
-
-        excludedRepos = set()
-        for nameWithOwner, repoData in self.repoDataByNameWithOwner.items():
-            for question in criteria:
-                # Handle repoType with default assumption
-                if question == "repoType":
-                    repoValue = repoData.get("repoType", (None, "Archival (intended for long-term maintenance)"))[1]
-                    if repoValue not in criteria["repoType"]:
-                        excludedRepos.add(nameWithOwner)
-                    continue
-
-                if question == "subjectType":
-                    # if subjectType is not present, assume "Biological specimen"
-                    repoValue = repoData.get("subjectType", (None, "Biological specimen"))[1]
-                    if repoValue not in criteria["subjectType"]:
-                        excludedRepos.add(nameWithOwner)
-                    continue
-
-                # Handle other criteria
-                if question in repoData:
-                    repoValue = repoData[question][1]
-                    if isinstance(repoValue, list):
-                        # Exclude only if NONE of the repo's values matches the criteria — decided
-                        # AFTER scanning all values.  (The old in-loop check excluded on the first
-                        # non-match, and a set can't un-exclude a later match.)  An empty list matches
-                        # nothing the user selected, so it is excluded.
-                        if not any(value in criteria[question] for value in repoValue):
-                            excludedRepos.add(nameWithOwner)
-                    else:
-                        if repoValue != "" and repoValue not in criteria[question]:
-                            excludedRepos.add(nameWithOwner)
-
-        matchString = f"*{criteria['freeText'].lower()}*"
-        matchingRepos = set()
-        textFields = ["githubRepoName", "species"]
-        if "Other" in criteria.get("subjectType", []):
-            textFields.append("otherSubjectDescription")
-        for nameWithOwner, repoData in self.repoDataByNameWithOwner.items():
-            if fnmatch.fnmatch(nameWithOwner, matchString):
-                matchingRepos.add(nameWithOwner)
-            for textField in textFields:
-                if textField in repoData:
-                    if fnmatch.fnmatch(repoData[textField][1].lower(), matchString):
-                        matchingRepos.add(nameWithOwner)
-
-        results = {}
-        for nameWithOwner in matchingRepos:
-            if nameWithOwner not in excludedRepos:
-                results[nameWithOwner] = self.repoDataByNameWithOwner[nameWithOwner]
-
-        return results
-
 
 class MorphoDepotTest(ScriptedLoadableModuleTest):
     """
