+++ MorphoDepot/MorphoDepotLib/logic_deps.py
@@ -1,0 +1,99 @@
+"""MorphoDepotLogic DepsMixin (split from MorphoDepot.py)."""
+import os
+import re
+import sys
+import csv
+import glob
+import json
+import time
+import math
+import locale
+import random
+import shutil
+import logging
+import platform
+import datetime
+import fnmatch
+import tempfile
+import traceback
+import subprocess
+from contextlib import contextmanager
+import git
+import requests
+import qt
+import ctk
+import slicer
+from slicer.i18n import tr as _
+from slicer.i18n import translate
+
+
+class DepsMixin:
+    def slicerVersionCheck(self):
+        return hasattr(slicer.vtkSegment, "SetTerminology")
+
+    def checkPythonDependencies(self):
+        """See if pygbif and idigbio are available.
+        The GitPython package is installed by default in slicer.
+        """
+        try:
+            import pygbif
+        except ModuleNotFoundError:
+            return False
+
+        try:
+            import idigbio
+        except ModuleNotFoundError:
+            return False
+
+        return True
+
+    def installPythonDependencies(self):
+        """Install pygbif and idigbio if needed
+        """
+        try:
+            import pygbif
+        except ModuleNotFoundError:
+            self.progressMethod(f"Installing pygbif")
+            slicer.util.pip_install("pygbif")
+            import pygbif
+
+        try:
+            import idigbio
+        except ModuleNotFoundError:
+            self.progressMethod(f"Installing idigbio")
+            slicer.util.pip_install("idigbio")
+            import idigbio
+
+    def checkCommand(self, command):
+        try:
+            completedProcess = subprocess.run(command, capture_output=True)
+            returnCode = completedProcess.returncode
+            stdout = completedProcess.stdout
+            stderr = completedProcess.stderr
+        except Exception as e:
+            stdout =  ""
+            stderr = str(e)
+            returnCode = -1
+        if returnCode != 0:
+            self.progressMethod(f"{command} failed to run, returned {returnCode}")
+            self.progressMethod(stdout)
+            self.progressMethod(stderr)
+            return False
+        return True
+
+    def checkGitDependencies(self):
+        """Check that git, and gh are available
+        """
+        if not (self.gitExecutablePath and self.ghExecutablePath):
+            self.progressMethod("git/gh paths are not set")
+            return False
+        if not (os.path.exists(self.gitExecutablePath) and os.path.exists(self.ghExecutablePath)):
+            self.progressMethod("bad git/gh paths")
+            self.progressMethod(f"git path is {self.gitExecutablePath}")
+            self.progressMethod(f"gh path is {self.ghExecutablePath}")
+            return False
+        if not self.checkCommand([self.gitExecutablePath, '--version']):
+            return False
+        if not self.checkCommand([self.ghExecutablePath, 'auth', 'status']):
+            return False
+        return True
